==> notebooks/7_make_figures.ipynb <==
# Make figures

import matplotlib.pyplot as plt
import matplotlib
from mpl_toolkits.axes_grid1.inset_locator import inset_axes
import pandas as pd
import geopandas as gpd
import numpy as np
import os
import glob
import sys
import seaborn as sns
from tqdm import tqdm
import rioxarray as rxr
import xarray as xr
import joblib
from scipy.stats import median_abs_deviation
import warnings
warnings.filterwarnings('ignore')

# Define path to this repository and import functions
code_path = '/Users/raineyaberle/Research/PhD/snow_cover_mapping/glacier-snow-cover-analysis/'
sys.path.append(os.path.join(code_path, 'scripts'))
import utils as f

# Define input and output paths
figures_out_path = os.path.join(code_path, 'figures')
scm_path = '/Volumes/LaCie/raineyaberle/Research/PhD/snow_cover_mapping/'
out_path = os.path.join(scm_path, 'analysis')

# Define font and font size
fontsize=12
plt.rcParams.update({'font.size':fontsize, 'font.sans-serif':'Arial'})

# Define clusters and AOIs file names
clusters_fn = os.path.join(out_path, 'climate_clusters.csv')
aois_fn = os.path.join(out_path, 'AOIs.gpkg')

# Define function for saving figure
def save_figure(fig, fig_fn):
    fig.savefig(fig_fn, dpi=300, bbox_inches='tight')
    print('Figure saved to file:', fig_fn)

## Define colormaps and order of clusters and subregions for plotting

# Climate clusters
cluster_cmap_dict = {'W. Aleutians': '#dd3497', 
                     'Maritime': '#018571',
                     'Transitional-Maritime': '#80cdc1',
                     'Transitional-Continental': '#dfc27d',
                     'Continental': '#a6611a',
                     }

cluster_order = ['W. Aleutians', 'Maritime', 'Transitional-Maritime', 'Transitional-Continental', 'Continental']
subregion_order = ['N. Rockies', 'Alaska Range', 'W. Chugach Mtns.', 'St. Elias Mtns.', 'N. Coast Ranges',
                   'Aleutians', 'N. Cascades', 'C. Rockies', 'S. Cascades', 'S. Rockies']

# Count total number of observations
rgi_ids = [os.path.basename(x) for x in sorted(glob.glob(os.path.join(scm_path, 'study-sites', 'RGI*')))]
nobs = np.zeros(len(rgi_ids))
for i, rgi_id in enumerate(rgi_ids):
    im_classified_fns = (glob.glob(os.path.join(scm_path, 'study-sites', rgi_id, 'classified', '20*.nc'))
                         + glob.glob(os.path.join(scm_path, 'study-sites', rgi_id, 'imagery', 'classified', '20*.nc')))
    if len(im_classified_fns) < 1:
        print(rgi_id)
    nobs[i] = len(im_classified_fns)

print(f'Total number of observations = {np.sum(nobs)}')
print(f'Mean observations per site = {np.mean(nobs)}')
print(f'Median observations per site = {np.median(nobs)}')
print(f'Minimum observations per site = {np.min(nobs)}')
print(f'Maximum observations per site = {np.max(nobs)}')


## Figure 1. Study sites and climate clusters

# Load climate clusters / mean climate
clusters = pd.read_csv(clusters_fn)
print('Climate clusters loaded')

# Load AOIs
aois = gpd.read_file(aois_fn)
# Add climate clusters column
aois = pd.merge(aois, clusters[['RGIId', 'clustName']], on='RGIId')
print('AOIs loaded')

# Load RGI O2 Regions
rgi_O2_fn = os.path.join(scm_path, '..', 'GIS_data', 'RGI', 'RGIv7_02Regions', 
                                'RGI2000-v7.0-o2regions-Alaska-westernCanadaUS_clipped_to_country_outlines.shp')
rgi_O2 = gpd.read_file(rgi_O2_fn)
# remove Brooks Range
rgi_O2 = rgi_O2.loc[rgi_O2['o2region']!='01-01']
# add subregion name and color column
rgi_O2[['Subregion', 'color']] = '', ''
for i, o1o2 in enumerate(rgi_O2['o2region'].drop_duplicates().values):
    o1 = int(o1o2[0:2])
    o2 = int(o1o2[3:])
    subregion_name, color = f.determine_subregion_name_color(o1, o2)
    rgi_O2.loc[rgi_O2['o2region']==o1o2, 'Subregion'] = subregion_name
print('RGI O2 regions loaded')

# Load GTOPO30
gtopo_fn = '/Users/raineyaberle/Research/PhD/GIS_data/GTOPO30_clip.tif'
gtopo = rxr.open_rasterio(gtopo_fn)
gtopo = xr.where(gtopo==-32768, np.nan, gtopo)
print('GTOPO30 loaded')

# Set up figure
plot_inset = True # whether to plot mean climate conditions inset
fig = plt.figure(figsize=(10,10))
if plot_inset:
    ax = [fig.add_subplot(1,1,1),
        fig.add_axes([0.82, 0.58, 0.07, 0.07]),
        fig.add_axes([0.26, 0.24, 0.27, 0.23])]
    legend_position = [0.7, 0.46, 0.2, 0.2]
else:
    ax = [fig.add_subplot(1,1,1),
          fig.add_axes([0.41, 0.31, 0.08, 0.08]),
          None]
    legend_position = [0.3, 0.2, 0.2, 0.2]    

### a) Map view
# GTOPO hillshade
ls = matplotlib.colors.LightSource(azdeg=90, altdeg=45)
ax[0].imshow(ls.hillshade(gtopo.data[0], vert_exag=0.002), cmap='gray', alpha=0.5,
             extent=(np.min(gtopo.x.data), np.max(gtopo.x.data), 
                     np.min(gtopo.y.data), np.max(gtopo.y.data)))
# RGI O2 region outlines
color = '#525252'
rgi_O2.plot(ax=ax[0], alpha=1.0, facecolor='None', edgecolor=color, linewidth=1)
ax[0].set_yticks(np.linspace(45, 65, num=6))
ax[0].set_xlim(-167, -112)
ax[0].set_ylim(46, 66)
ax[0].set_xlabel('Longitude ($^{\circ}$E)')
ax[0].set_ylabel('Latitude ($^{\circ}$N)')
ax[0].set_aspect(2.2)
# Site locations
sns.scatterplot(data=aois, x='CenLon', y='CenLat', edgecolor='k', linewidth=0.5, s=20,
                hue='clustName', palette=cluster_cmap_dict, hue_order=cluster_order, alpha=1, ax=ax[0])
handles, labels = ax[0].get_legend_handles_labels()
ax[0].legend().remove()
# Add region labels and arrows
fontweight = 'bold'
background_color = [1, 1, 1, 0.5]
ax[0].text(-160.5, 56.5, f"Aleutians\n(N={len(aois.loc[aois['Subregion']=='Aleutians'])})", ha='center', color=color, rotation=35, fontsize=fontsize-3, fontweight=fontweight)
ax[0].text(-157, 62.5, f"Alaska Range\n(N={len(aois.loc[aois['Subregion']=='Alaska Range'])})", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].text(-147.7, 57.4, f"W. Chugach \nMtns.\n(N={len(aois.loc[aois['Subregion']=='W. Chugach Mtns.'])})", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].arrow(-147.6, 58.9, 0, 0.6, color=color, linewidth=2, head_width=0.34, head_length=0.2)
ax[0].text(-141.5, 57.4, f"St. Elias \nMtns.\n(N={len(aois.loc[aois['Subregion']=='St. Elias Mtns.'])})", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].arrow(-141.5, 58.9, 0, 0.6, color=color, linewidth=2, head_width=0.34, head_length=0.2)
ax[0].text(-139.8, 55.5, f"N. Coast \nRanges\n(N={len(aois.loc[aois['Subregion']=='N. Coast Ranges'])})", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].arrow(-137.1, 56.5, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[0].text(-133.4, 50.9, f"N. Cascades\n(N={len(aois.loc[aois['Subregion']=='N. Cascades'])})", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].arrow(-129.9, 51.4, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[0].text(-129.7, 47, f"S. Cascades\n(N={len(aois.loc[aois['Subregion']=='S. Cascades'])})", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].arrow(-126.3, 47.5, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[0].text(-129.5, 64.5, f"N. Rockies\n(N={len(aois.loc[aois['Subregion']=='N. Rockies'])})", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].text(-119, 54.5, f"C. Rockies\n(N={len(aois.loc[aois['Subregion']=='C. Rockies'])})", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[0].text(-115, 47, f"S. Rockies\n(N={len(aois.loc[aois['Subregion']=='S. Rockies'])})", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)

# stars for the USGS Benchmark Glaciers
bbox_dict = dict(boxstyle='round,pad=0.3', linewidth=0.2, edgecolor='k', facecolor='w', alpha=0.5)
arrow_dict = dict(arrowstyle='->', color='k', connectionstyle="arc3,rad=0")
for name in ['Gulkana Glacier', 'Wolverine Glacier', 'Lemon Creek Glacier', 'Sperry Glacier MT', 'South Cascade Glacier WA']:
    aoi = aois.loc[aois['Name']==name]
    centroid = (aoi['CenLon'].values[0], aoi['CenLat'].values[0])
    ax[0].plot(*centroid, '*', markerfacecolor=cluster_cmap_dict[aoi['clustName'].values[0]], markeredgecolor='k', markeredgewidth=1, markersize=10)
    if ('Gulkana' in name) | ('Wolverine' in name) | ('Sperry' in name):
        x1, y1, x2, y2, ha = centroid[0], centroid[1]+1.5, centroid[0], centroid[1]+0.1, 'center'
    elif 'Lemon' in name:
        x1, y1, x2, y2, ha = centroid[0]+1.5, centroid[1], centroid[0]+0.1, centroid[1], 'left'
    elif 'South' in name:
        x1, y1, x2, y2, ha = centroid[0]+2, centroid[1]+0.3, centroid[0]+0.1, centroid[1], 'left'
    ax[0].text(x1, y1, name.replace(' MT','').replace(' WA','').replace(' ','\n'), color='k', fontsize=fontsize-5, ha=ha, bbox=bbox_dict)
    ax[0].annotate("", xy=(x2, y2), xytext=(x1, y1), arrowprops=arrow_dict)


# Legend
legend = fig.legend(handles, labels, loc='upper right', title='Climate cluster', bbox_to_anchor=legend_position, 
                    fontsize=fontsize-1, markerscale=2, alignment='left', labelspacing=0.6, framealpha=1)

# pie chart inset
pie_labels = cluster_order
pie_sizes = [len(clusters.loc[clusters['clustName']==label]) for label in pie_labels]
ax[1].pie(pie_sizes, colors=list(cluster_cmap_dict.values()), autopct=lambda p: f'{p * sum(pie_sizes) / 100:.0f}', 
        wedgeprops={'linewidth': 0.5, 'edgecolor': 'k'}, textprops={'fontsize': 7, 'fontweight': 'bold', 'color': 'k'},
        pctdistance=0.8)
ax[1].set_zorder(legend.get_zorder() + 1)

### b) Mean weather conditions
if plot_inset:
    sns.scatterplot(data=clusters, x='mean_annual_temp_range', y='mean_annual_precip_cumsum', s=20,
                    edgecolor='k', linewidth=0.5, hue='clustName', palette=cluster_cmap_dict, alpha=1, 
                    hue_order=cluster_order, legend=False, ax=ax[2])
    ax[2].set_xlabel('Air temperature range [$^{\circ}$C]', fontsize=fontsize-2)
    ax[2].set_ylabel('Precipitation sum [m]', fontsize=fontsize-2)
    ax[2].tick_params(labelsize=fontsize-2)

# Add text labels
if plot_inset:
    text_labels = ['a', 'b']
    for i, axis in enumerate([ax[0], ax[2]]):
        if i==1:
            scale = 0.85
        else:
            scale = 0.93
        axis.text((axis.get_xlim()[1] - axis.get_xlim()[0])*scale + axis.get_xlim()[0],
                (axis.get_ylim()[1] - axis.get_ylim()[0])*scale + axis.get_ylim()[0],
                text_labels[i], fontsize=fontsize+4, fontweight='bold')

# fig.tight_layout()
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'fig01_study_sites_clusters.png')
save_figure(fig, fig_fn)


## Figure 2. Median AARs and timing comparisons

# -----Load glacier boundaries
aois = gpd.read_file(aois_fn)
print('Glacier boundaries loaded')

# -----Load climate clusters
clusters = pd.read_csv(clusters_fn)

# -----Load median AARs for all sites
min_scs_fn = os.path.join(out_path, 'minimum_snow_cover_stats.csv')
min_scs = pd.read_csv(min_scs_fn)
# Add difference from September AAR
# min_scs['AAR_P50_diff'] = min_scs['AAR_P50_WOY39'] - min_scs['AAR_P50_min']
# Add Subregion and climate cluster info
min_scs[['CenLon', 'CenLat', 'Subregion', 'clustName']] = 0, 0, '', ''
for rgi_id in min_scs['RGIId'].drop_duplicates().values:
    cenlon, cenlat, subregion = aois.loc[aois['RGIId']==rgi_id, ['CenLon', 'CenLat', 'Subregion']].values[0]
    cluster = clusters.loc[clusters['RGIId']==rgi_id, 'clustName'].values[0]
    min_scs.loc[min_scs['RGIId']==rgi_id, ['CenLon', 'CenLat', 'Subregion', 'clustName']] = cenlon, cenlat, subregion, cluster
# Sort by subregion order
min_scs['Subregion'] = pd.Categorical(min_scs['Subregion'], subregion_order)
min_scs.sort_values(by='Subregion', inplace=True)
print('Median AARs loaded')

# -----Load melt season timings estimate
melt_season_fn = os.path.join(out_path, 'melt_season_timing.csv')
melt_season = pd.read_csv(melt_season_fn)
# Add subregion column
melt_season = pd.merge(melt_season, aois[['RGIId', 'Subregion']], on='RGIId')
# Sort by subregion order
melt_season['Subregion'] = pd.Categorical(melt_season['Subregion'], subregion_order)
melt_season.sort_values(by='Subregion', inplace=True)
print('Melt season timing loaded')

lw = 1.0
gs = matplotlib.gridspec.GridSpec(10,2, wspace=0.0, hspace=0.0)
fig = plt.figure(figsize=(8,8))
ax = []

# Iterate over subregions
median_color = 'w'
fill_color = '#4f97d6' #'#2171b5'
Iax = -1
for i, subregion in enumerate(min_scs['Subregion'].unique()):
    min_scs_subregion = min_scs.loc[min_scs['Subregion']==subregion]
    
    # a) AARs
    ax.append(fig.add_subplot(gs[i,0]))
    Iax += 1
    min_scs_subregion_melt = pd.melt(
            min_scs_subregion,
            id_vars=['RGIId', 'Subregion'],
            value_vars=['AAR_median'],
            var_name='AAR_type',
            value_name='blank'
    )
    min_scs_subregion_melt.rename(columns={'blank': 'AAR'}, inplace=True)
    sns.violinplot(data=min_scs_subregion, x='AAR_median', color=fill_color, width=0.8, 
                   inner='quart', ax=ax[Iax], linecolor='k', cut=0, legend=False)
    ax[Iax].lines[1].set_color('w')
    ax[Iax].lines[1].set_linestyle('-')
    ax[Iax].set_xlim(0,1)
    ax[Iax].spines[['right', 'top']].set_visible(False)
    ax[Iax].set_ylabel('')
    if i > 0:
        ax[Iax].spines['top'].set_visible(True)
        ax[Iax].spines['top'].set_color('gray')
    if i==0:
        ax[Iax].text(0.8, 0.7, 'a', transform=ax[Iax].transAxes, 
                     fontweight='bold', fontsize=fontsize+4, color='k')
    if i < 9:
        ax[Iax].set_xticks([])
        ax[Iax].set_xlabel('')
        ax[Iax].spines['bottom'].set_color('gray')
    else:
        ax[Iax].set_xlabel('Accumulation area ratio')
    ax[Iax].set_ylabel(subregion, ha='right', va='center', color='k', rotation=0)

    # b) AAR timing and melt season duration
    ax.append(fig.add_subplot(gs[i,1]))
    Iax += 1
    k = sns.kdeplot(min_scs_subregion['WOY_median'], vertical=False, color=fill_color, 
                    fill=True, edgecolor='k', linewidth=lw, alpha=1, ax=ax[Iax], zorder=2)
    median = min_scs_subregion['WOY_median'].median()
    ax[Iax].plot([median, median], [0, ax[Iax].get_ylim()[1]*0.9], '-', color='w', linewidth=lw+0.5, zorder=3)
    ax[Iax].plot([14,45], [0,0], '-', color='k', linewidth=2)
    melt_season_subregion = melt_season.loc[melt_season['Subregion']==subregion]
    melt_season_start = melt_season_subregion['melt_season_start_WOY'].mean()
    melt_season_end = melt_season_subregion['melt_season_end_WOY'].mean()
    ax[Iax].fill_between([melt_season_start, melt_season_end],  
                         [0, 0], [ax[Iax].get_ylim()[1], ax[Iax].get_ylim()[1]],
                         facecolor='k', edgecolor='None', alpha=0.15, zorder=1, label='Melt season duration')
    ax[Iax].set_xlim(13,45)
    ax[Iax].set_yticks([])
    ax[Iax].set_ylabel('')
    ax[Iax].set_xticks([])
    ax[Iax].set_xlabel('')
    if i==9:
        ax[Iax].set_xticks([18, 22, 26, 31, 35, 39, 44])
        ax[Iax].set_xticklabels(['May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov'])
        ax[Iax].set_xlabel('Snow minimum timing')
    if i==0:
        ax[Iax].legend(loc='upper center', frameon=False, bbox_to_anchor=[0.4, 1.6, 0.2, 0.2])
        ax[Iax].text(0.8, 0.7, 'b', transform=ax[Iax].transAxes, 
                     fontweight='bold', fontsize=fontsize+4, color='k')
    ax[Iax].spines[['left', 'right', 'top', 'bottom']].set_visible(False)
            
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'fig02_median_aars+timings.png')
save_figure(fig, fig_fn)

# Plot AAR as a function of terrain characteristics

from sklearn.linear_model import LinearRegression

plt.rcParams.update({'font.sans-serif': 'Arial', 'font.size': 12})
subregions_cmap_dict = {'N. Rockies': '#b15928',
 'Alaska Range': '#6a3d9a',
 'W. Chugach Mtns.': '#cab2d6',
 'St. Elias Mtns.': '#33a02c',
 'N. Coast Ranges': '#b2df8a',
 'Aleutians': '#fb9a99',
 'N. Cascades': '#1f78b4',
 'C. Rockies': '#ff7f00',
 'S. Cascades': '#a6cee3',
 'S. Rockies': '#fdbf6f'}

if 'Area' not in min_scs.keys():
    min_scs = pd.merge(min_scs, aois[['RGIId', 'Area', 'Zmin', 'Zmax', 'Aspect', 'Slope']], on='RGIId')
    min_scs['Zrange'] = min_scs['Zmax'] - min_scs['Zmin']
cols = ['Area', 'Zrange', 'Slope', 'Aspect']
labels = ['Area [km$^2$]', 'Elevation range [m]', 'Slope [degrees]', 'Aspect']

fig, ax = plt.subplots(2, 2, figsize=(8,8))
ax = ax.flatten()
for i, (col, label) in enumerate(zip(cols, labels)):
    # distribution
    sns.scatterplot(min_scs, x=col, y='AAR_median', hue='Subregion', hue_order=subregion_order, palette=subregions_cmap_dict, ax=ax[i])
    # add legend on first iteration
    if i==0:
        handles, labels = ax[i].get_legend_handles_labels()
        fig.legend(handles, labels, loc='center right', bbox_to_anchor=[1.05, 0.4, 0.2, 0.2])
        ax[i].set_xscale('log')
    ax[i].legend().remove()
    if i==len(cols)-1:
        ax[i].set_xticks([0, 90, 180, 270])
        ax[i].set_xticklabels(['N', 'E', 'S', 'W'])
    # linear fit
    fit = LinearRegression().fit(min_scs[col].values.reshape(-1,1), min_scs['AAR_median'])
    x = np.arange(min_scs[col].min(), min_scs[col].max()).reshape(-1,1)
    yfit = fit.predict(x)
    r2 = fit.score(min_scs[col].values.reshape(-1,1), min_scs['AAR_median'])
    ax[i].plot(x, yfit, '-k')
    if i==0:
        ax[i].text(100, 0.94, f"R$^2$ = {np.round(r2,3)}", ha='right')
    else:
        ax[i].text(np.nanmean(x), 0.94, f"R$^2$ = {np.round(r2,3)}", ha='center')
    ax[i].set_ylim(-0.05, 1.0)
    ax[i].set_xlabel(label)
    ax[i].set_ylabel('AAR')
    ax[i].grid()
    fig.tight_layout()
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'AAR_vs_area_elevation.png')
save_figure(fig, fig_fn)


# Check WOY and months
df = pd.DataFrame({'Date': pd.date_range('2013-01-01', '2023-12-30')})
df['WOY'] = df['Date'].dt.isocalendar().week
df['year'] = df['Date'].dt.year
df['month'] = df['Date'].dt.month
df['day'] = df['Date'].dt.day
# df.loc[(df['month']==10) & (df['day']==1)]['WOY']#.median()
df.loc[df['WOY']==38]

# Print statistics

print('All sites:')
print('-----------')
print('---AARs---')
print(min_scs['AAR_median'].describe())
print('\n---WOYs---')
print(min_scs['WOY_median'].describe()) 

print(' ')
print('By subregion:')
print('-----------')
print('---AARs---')
print(min_scs.groupby('Subregion')['AAR_median'].describe().sort_values(by='50%'))

print('\n---WOYs---')
print(min_scs.groupby('Subregion')['WOY_median'].describe().sort_values(by='50%'))


# Look at which sites had AARs < 0.1
aar_lt_p1 = min_scs.loc[min_scs['AAR_median'] < 0.1]
print(len(aar_lt_p1))
aar_lt_p1


scs_MCs_fn = os.path.join(out_path, 'median_snow_cover_stats_MC.nc')
scs_MCs = xr.open_dataset(scs_MCs_fn)

# AARs
min_aar = scs_MCs['AAR'].min(dim='WOY')
min_aar_ranges = [np.nanmax(min_aar.isel(RGIId=i).values) - np.nanmin(min_aar.isel(RGIId=i).values) for i in np.arange(len(min_aar.RGIId))]
min_aar_mads = [median_abs_deviation(min_aar.isel(RGIId=i).values) for i in np.arange(len(min_aar.RGIId))]

# Print results
print("AAR:")
print(f"Minimum range in AARs across all sites and simulations:", np.nanmin(min_aar_ranges))
print(f"Maximum range in AARs across all sites and simulations:", np.nanmax(min_aar_ranges))
print(f"Minimum MAD in AARs across simulations and sites:", np.nanmin(min_aar_mads))
print(f"Maximum MAD in AARs across simulations and sites:", np.nanmax(min_aar_mads))
print(f"Mean MAD in AARs across simulations and sites:", np.nanmean(min_aar_mads))
print('Mean range in AARs across sites and simulations:', np.nanmean(min_aar_ranges))

# Snow minima timings
min_aar_woy = scs_MCs['WOY'].isel(WOY=scs_MCs['AAR'].argmin(dim='WOY'))
min_aar_woy_ranges = [np.nanmax(min_aar_woy.isel(RGIId=i).values) - np.nanmin(min_aar_woy.isel(RGIId=i).values) for i in np.arange(len(min_aar_woy.RGIId))]
min_aar_woy_mads = [median_abs_deviation(min_aar_woy.isel(RGIId=i).values) for i in np.arange(len(min_aar_woy.RGIId))]

# Print results
print("\nSnow minima timings:")
print(f"Minimum range in WOYs across all sites and simulations:", np.nanmin(min_aar_woy_ranges))
print(f"Maximum range in WOYs across all sites and simulations:", np.nanmax(min_aar_woy_ranges))

print(f"Minimum MAD in WOYs across simulations and sites:", np.nanmin(min_aar_woy_mads))
print(f"Maximum MAD in WOYs across simulations and sites:", np.nanmax(min_aar_woy_mads))
print(f"Mean MAD in WOYs across simulations and sites:", np.nanmean(min_aar_woy_mads))
print('Mean range in WOYs across sites and simulations:', np.nanmean(min_aar_woy_ranges))

# # Print results
# print("\nSnow minima timings:")
# print(f"Mean WOY range across sites: {mean_range_woy.values}")
# print(f"Mean WOY IQR across sites: {mean_iqr_woy.values}")
# print(f"Mean WOY standard deviation across sites: {mean_std_dev_woy.values}")


## Figure 3. Observed vs. modeled SMB

# Load AOIs
aois = gpd.read_file(aois_fn)
aois[['O1Region', 'O2Region']] = aois[['O1Region', 'O2Region']].astype(int)
print('Compiled AOIs loaded')

# Load climate clusters
clusters_fn = os.path.join(out_path, 'climate_clusters.csv')
clusters = pd.read_csv(clusters_fn)
print('Clusters loaded')

# Load merged monthly SLAs, ELAs, 
slas_elas_merged_fn = os.path.join(out_path, 'monthly_SLAs_annual_ELAs_observed_modeled.nc')
slas_elas_merged = xr.load_dataset(slas_elas_merged_fn)
# Convert to pandas dataframes for compatibility with seaborn
slas = slas_elas_merged[['SLA_obs', 'SLA_mod', 'SMB_at_SLA_obs']].to_dataframe().reset_index()
slas['Month'] = slas['time'].dt.month
elas = slas_elas_merged[['ELA_obs', 'ELA_mod']].to_dataframe().reset_index()
print('SLAs and ELAs loaded')

# Load PyGEM comparisons
pygem_params_fn = os.path.join(out_path, 'PyGEM_comparison_params.csv')
pygem_params = pd.read_csv(pygem_params_fn)
print('PyGEM parameter comparisons loaded')

# Add subregion, centroid coordinates, and climate cluster columns
def merge_sort_df(df):
    df = df.merge(clusters[['RGIId', 'clustName']], on='RGIId')
    df = df.merge(aois[['RGIId', 'CenLon', 'CenLat', 'Subregion']], on='RGIId')
    df['clustName'] = pd.Categorical(df['clustName'], categories=cluster_order, ordered=True)
    df.sort_values(by=['clustName', 'RGIId'], inplace=True)
    return df
slas = merge_sort_df(slas)
elas = merge_sort_df(elas)
pygem_params = merge_sort_df(pygem_params)
print('Added columns for plotting')

# Plot
fig, ax = plt.subplots(2, 2, figsize=(10,8), gridspec_kw=dict(width_ratios=[1.5,1]))
ax = ax.flatten()

### a) Monthly snowline differences
slas['SLA_mod-obs'] = slas['SLA_mod'] - slas['SLA_obs']
sns.boxplot(data=slas, x='Month', y='SLA_mod-obs', hue='clustName', hue_order=cluster_order,
            palette=cluster_cmap_dict, saturation=1, showfliers=False,
            boxprops=dict(linewidth=1, edgecolor='k'), whiskerprops=dict(linewidth=1, color='k'), 
            ax=ax[0])
ax[0].set_xticks(np.arange(0,5))
ax[0].set_xticklabels(['May', 'Jun', 'Jul', 'Aug', 'Sep'])
ax[0].set_xlim(-0.5, 4.5)
ax[0].set_ylim(-550, 1300)
ax[0].set_title('a) Modeled $-$ observed snowline altitudes')
ax[0].set_ylabel('Snowline altitude difference [m]')
# add minor grid lines
ax[0].xaxis.set_minor_locator(matplotlib.ticker.MultipleLocator(0.5))
ax[0].grid(which='minor')
# remove minor ticks
ax[0].tick_params(axis='x', which='minor', bottom=False)

### b) ELA differences
elas['ELA_mod-obs'] = elas['ELA_mod'] - elas['ELA_obs']
sns.kdeplot(data=elas, y='ELA_mod-obs', hue='clustName', hue_order=cluster_order, fill=True, alpha=0.1,
            linewidth=2, palette=cluster_cmap_dict, ax=ax[1])
ax[1].set_ylim(ax[0].get_ylim())
ax[1].set_title('b) Modeled $-$ observed ELAs')
ax[1].set_xlabel('Relative density')
ax[1].set_ylabel('ELA difference [m]')
ax[1].set_xticks([])

### c) Modeled SMB at remotely-sensed snowlines
sns.boxplot(data=slas, x='Month', y='SMB_at_SLA_obs', hue='clustName', hue_order=cluster_order,
            palette=cluster_cmap_dict, saturation=1, showfliers=False, legend=False,
            boxprops=dict(linewidth=1, edgecolor='k'), whiskerprops=dict(linewidth=1, color='k'), 
            ax=ax[2])
ax[2].set_xlabel('Month')
ax[2].set_xticks(np.arange(0,5))
ax[2].set_xticklabels(['May', 'Jun', 'Jul', 'Aug', 'Sep'])
ax[2].set_xlim(-0.5, 4.5)
ax[2].set_ylabel('SMB [m.w.e.]')
ax[2].set_title('c) Modeled SMB at observed snowline altitude')
# add minor grid lines
ax[2].xaxis.set_minor_locator(matplotlib.ticker.MultipleLocator(0.5))
ax[2].grid(which='minor')
# remove minor ticks
ax[2].tick_params(axis='x', which='minor', bottom=False)

### d) Original vs. adjusted PyGEM parameters
def add_arrow(axis, x1, y1, x2, y2, arrowstyle='-|>', color='black'):
    axis.annotate("", xy=(x2, y2), xytext=(x1, y1),
                  arrowprops=dict(arrowstyle=arrowstyle, color=color,
                                  connectionstyle="arc3,rad=0"))

for rgi_id in pygem_params['RGIId'].drop_duplicates().values:
    pygem_site = pygem_params.loc[pygem_params['RGIId']==rgi_id]
    color = cluster_cmap_dict[pygem_site['clustName'].values[0]]
    # line w/ arrow between points
    add_arrow(ax[3], 
              pygem_site.loc[pygem_site['parameter']=='tbias', 'Original'].values[0],
              pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Original'].values[0] * 1e3,
              pygem_site.loc[pygem_site['parameter']=='tbias', 'Best'].values[0],
              pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Best'].values[0] * 1e3,
              color=color)
    # original points
    ax[3].plot(pygem_site.loc[pygem_site['parameter']=='tbias', 'Original'].values[0],
               pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Original'].values[0] * 1e3, 
               'o', color=color,
               markersize=pygem_site.loc[pygem_site['parameter']=='kp', 'Original'].values[0]*3)
    # best points
    ax[3].plot(pygem_site.loc[pygem_site['parameter']=='tbias', 'Best'].values[0],
               pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Best'].values[0] * 1e3, 
               'o', color=color,
               markersize=pygem_site.loc[pygem_site['parameter']=='kp', 'Best'].values[0]*3)
    # Name annotation
    if (pygem_site['Name'].values[0]=='South Cascade'):
        ax[3].annotate(pygem_site['Name'].values[0].replace(' ','\n'), 
                    xy=(pygem_site.loc[pygem_site['parameter']=='tbias', 'Best'].values[0] - 0.1, 
                        pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Best'].values[0] * 1e3 - 0.35),
                        color='k', ha='left', fontsize=10)
    else:
        ax[3].annotate(pygem_site['Name'].values[0].replace(' ','\n'), 
                    xy=(pygem_site.loc[pygem_site['parameter']=='tbias', 'Best'].values[0] + 0.2, 
                        pygem_site.loc[pygem_site['parameter']=='ddfsnow', 'Best'].values[0] * 1e3 - 0.1),
                        color='k', ha='left', fontsize=10)
    # dummy points for legend
    if pygem_site['Name'].values[0]=='Gulkana':
        for kp in [1, 2, 3]:
            ax[3].plot(-100, -100, 'ok', markersize=kp*3, label=str(kp))
    
ax[3].set_xlim(-4,4.6)
ax[3].set_xlabel('Temperature bias [$^{\circ}$C]')
ax[3].set_ylim(1.3, 4.7)
ax[3].set_ylabel('Degree-day factor of snow [m $^{\circ}$C$^{-1}$ d$^{-1}$]')
ax[3].set_title('d) Original vs. adjusted PyGEM parameters')
ax[3].legend(loc='lower left', ncol=3, handletextpad=0.2, title='Precipitation factor')

# Add legend
handles, labels = ax[0].get_legend_handles_labels()
ax[0].legend().remove()
ax[1].legend().remove()
fig.legend(handles, labels, loc='lower center', ncols=5, frameon=False,
           bbox_to_anchor=[0.43, -0.05, 0.2, 0.2], labelspacing=0.6, handletextpad=0.3)

# Add text label and line at 0
text_labels = ['a', 'b']
for i, axis in enumerate(ax[0:3]):
    axis.axhline(0, color='k')
    
# fig.subplots_adjust(wspace=0.2)
fig.tight_layout()
plt.show()

# Save to file
fig_fn = os.path.join(figures_out_path, 'fig03_modeled_observed_SMB_differences.png')
save_figure(fig, fig_fn)


# SLA stats
print(slas.groupby(['Month'])['SLA_mod-obs'].describe())
slas.groupby(['clustName', 'Month'])['SLA_mod-obs'].describe()

# ELA stats
print(elas['ELA_mod-obs'].describe())
elas.groupby('clustName')['ELA_mod-obs'].describe().sort_values(by='50%')

# PyGEM simulations
pygem_params.loc[pygem_params['parameter']=='ddfsnow']

## Figure S1. Sites distributions

# Load analyzed glacier boundaries
aois = gpd.read_file(aois_fn)
cols = ['O1Region', 'O2Region', 'Zmed', 'Aspect', 'Slope', 'Area']
for col in cols:
    aois[col] = aois[col].astype(float)
    
# Load all glacier boundaries in O1 regions 1 and 2
rgi_path = '/Volumes/LaCie/raineyaberle/Research/PhD/GIS_data/RGI/'
rgi_fns = ['01_rgi60_Alaska/01_rgi60_Alaska.shp',
           '02_rgi60_WesternCanadaUS/02_rgi60_WesternCanadaUS.shp']
rgi = gpd.GeoDataFrame()
for rgi_fn in rgi_fns:
    file = gpd.read_file(os.path.join(rgi_path, rgi_fn))
    rgi = pd.concat([rgi, file])
rgi[['O1Region', 'O2Region']] = rgi[['O1Region', 'O2Region']].astype(int)
# Add column for subregion name
rgi = pd.merge(rgi, aois[['O1Region', 'O2Region', 'Subregion']], on=['O1Region', 'O2Region'])

# Define plotting variables
columns = ['Zmed', 'Aspect', 'Slope', 'Area']
xlabels = ['Median elevation [m]', 'Aspect [degrees]', 'Slope [degrees]', 'Area [km$^2$]']
bins_list = [np.linspace(0, 361, num=20), # Aspect
             np.linspace(0, 51, num=20), # Slope
             np.linspace(0, 300, num=50)] # Area

# Set up figure
plt.rcParams.update({'font.sans-serif': 'Arial', 'font.size': 12})
fig, ax = plt.subplots(len(subregion_order)+1, 4, figsize=(12, (len(subregion_order)+1)*1.5))
aois_color = '#b35806'

# All subregions
for j, (column, xlabel) in enumerate(zip(columns, xlabels)):
    if column=='Zmed': 
        bins = np.linspace(rgi['Zmed'].min(), rgi['Zmed'].max(), num=20)
    else:
        bins  = bins_list[j-1]
    ax[0,j].hist(rgi[column].values, bins=bins, facecolor='k', alpha=0.6)
    ax[0,j].set_title(xlabel)
    ax2 = ax[0,j].twinx()
    ax2.hist(aois[column].values, bins=bins, facecolor=aois_color, alpha=0.6)
    ax2.set_yticks(ax2.get_yticks())
    ax2.set_yticklabels(ax2.get_yticklabels(), color=aois_color)
    ax2.spines['right'].set_color(aois_color)
    ax2.tick_params(axis='y', color=aois_color)
ax[0,0].set_ylabel('All regions', fontweight='bold')

# Individual subregions
for i, subregion in enumerate(subregion_order):
    # Subset glaciers
    aois_subregion = aois.loc[aois['Subregion']==subregion]
    rgi_subregion = rgi.loc[rgi['Subregion']==subregion]

    # Plot all glaciers in subregion
    for j, (column, xlabel) in enumerate(zip(columns, xlabels)):
        if column=='Zmed':
            bins = np.linspace(rgi_subregion['Zmed'].min(), rgi_subregion['Zmed'].max(), num=20)
        else:
            bins = bins_list[j-1]
        ax[i+1,j].hist(rgi_subregion[column].values, bins=bins, facecolor='k', alpha=0.6)
        if j==0:
            ax[i+1,j].set_ylabel(subregion)
        ax2 = ax[i+1,j].twinx()
        ax2.hist(aois_subregion[column].values, bins=bins, facecolor=aois_color, alpha=0.6)
        ax2.set_yticks(ax2.get_yticks())
        ax2.set_yticklabels(ax2.get_yticklabels(), color=aois_color)
        ax2.spines['right'].set_color(aois_color)
        ax2.tick_params(axis='y', color=aois_color)

fig.tight_layout()
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'figS1_site_distributions.png')
save_figure(fig, fig_fn)

len(rgi.loc[rgi['Area'] < 0.1]) / len(rgi)

## Figure S2. Snowline altitude uncertainty analysis

# Example site ID
rgi_id = 'RGI60-01.01104'
epsg_utm = "EPSG:32606"

# Load SLA bounds
sla_bounds_fn = os.path.join(out_path, 'SLA_uncertainty_analysis.nc')
sla_bounds = xr.open_dataset(sla_bounds_fn)
sla_bounds['SLA_bounds_range'] = sla_bounds['SLA_upper_bound'] - sla_bounds['SLA_lower_bound']

# Load all AOIs
aois = gpd.read_file(aois_fn)
# Add elevation range column
aois['Zrange'] = aois['Zmax'] - aois['Zmin']
# Merge with SLA bounds
sla_bounds['Zrange'] = (('RGIId'), [aois.loc[aois['RGIId']==rgi_id, 'Zrange'].values[0] 
                                    for rgi_id in sla_bounds['RGIId'].values])
# Load snow cover stats
scs_fn = os.path.join(scm_path, 'study-sites', rgi_id, f"{rgi_id}_classifications.zarr")
scs = xr.open_dataset(scs_fn)
# Load DEM
dem_fn = glob.glob(os.path.join(scm_path, 'study-sites', rgi_id, 'DEMs', '*.tif'))[0]
dem = rxr.open_rasterio(dem_fn).isel(band=0)
dem = dem.rio.write_crs("EPSG:4326")
# Load AOI
aoi_fn = os.path.join(scm_path, 'study-sites', rgi_id, 'AOIs', f"{rgi_id}_outline.shp")
aoi = gpd.read_file(aoi_fn)
# Clip DEM to AOI
dem = dem.rio.clip(aoi.geometry)
# Choose an image with some masking and relatively low transient AAR for demonstration
sc = (scs.sel(time=slice(pd.Timestamp('2019-07-03'), pd.Timestamp('2019-07-05'))).isel(time=0))
classified = sc.classification
classified = xr.where(classified==0, np.nan, classified)
classified = classified.rio.write_crs("EPSG:4326")
classified_utm = classified.rio.reproject(epsg_utm)
classified_utm = xr.where(classified_utm > 1e30, np.nan, classified_utm)
classified_utm = classified_utm.rio.write_crs(epsg_utm)

# Create binary snow image
snow_binary = xr.where((classified_utm==1) | (classified_utm==2), 1, 0)
snow_binary = xr.where(np.isnan(classified_utm), np.nan, snow_binary) # re-insert no data values

# Regrid DEM to classified image grid
dem_utm = dem.rio.reproject_match(classified_utm)
dem_utm = xr.where((dem_utm < -1e3) | (dem_utm > 1e4), np.nan, dem_utm)

# Load RGB image from GEE
import geedim as gd
import wxee
import ee
ee.Initialize(project='ee-raineyaberle')
aoi_ee = ee.Geometry.Polygon(list(zip(aoi.geometry[0].exterior.coords.xy[0], 
                                      aoi.geometry[0].exterior.coords.xy[1]))).buffer(1e3)
im_col_ee = gd.MaskedCollection.from_name("COPERNICUS/S2_SR_HARMONIZED").search(start_date='2019-07-03',
                                                                            end_date='2019-07-04',
                                                                            region=aoi_ee)
im_ee = im_col_ee.ee_collection.first()
im_ee = im_ee.select(['B2', 'B3', 'B4'])
rgb_im = im_ee.wx.to_xarray(region=aoi_ee, scale=10)
rgb_im = rgb_im.rio.reproject(epsg_utm)
rgb_im = xr.where(rgb_im==-32768, np.nan, rgb_im/1e4).isel(time=0)

# Calculate lower and upper bounds of snowline altitude
sla_percentile = 1-sc.AAR.values
sla = np.nanquantile(dem.data.ravel(), sla_percentile)
sla_upper_percentile = (xr.where((dem_utm > sla) & (classified_utm > 2), 1, 0).sum().values / 
                        xr.where(~np.isnan(dem_utm), 1, 0).sum().values)
sla_upper_bound = np.nanquantile(dem.data.ravel(), sla_percentile + sla_upper_percentile)
sla_lower_percentile = (xr.where((dem_utm < sla) & (classified_utm <= 2), 1, 0).sum().values / 
                        xr.where(~np.isnan(dem_utm), 1, 0).sum().values)
sla_lower_bound = np.nanquantile(dem.data.ravel(), sla_percentile - sla_lower_percentile)
dem_utm = xr.where(dem_utm > 1e30, np.nan, dem_utm)
# Define colormap for classified images
cmap_dict = {"Snow": "#4eb3d3",  "Shadowed_snow": "#4eb3d3", "Ice": "#084081", "Rock": "#fe9929", "Water": "#969696"}
colors = []
for key in list(cmap_dict.keys()):
    color = list(matplotlib.colors.to_rgb(cmap_dict[key]))
    if key=='Rock':
        color += [0.5]
    colors.append(color)
cmap = matplotlib.colors.ListedColormap(colors)
cmap

# Set up figure
fontsize=11
lw=1.5
plt.rcParams.update({'font.size':fontsize, 'font.sans-serif': 'Arial'})
gs = matplotlib.gridspec.GridSpec(3,3)#, height_ratios=[1, 1.5, 1.5])
fig = plt.figure(figsize=(8,12))
ax = [fig.add_subplot(gs[0,0]), fig.add_subplot(gs[0,1]), fig.add_subplot(gs[0,2]),
      fig.add_subplot(gs[1,:]),
      fig.add_subplot(gs[2,:])]

# DEM
im = ax[0].imshow(dem_utm.data, cmap='terrain', 
                  extent=(np.min(dem_utm.x)/1e3, np.max(dem_utm.x)/1e3, np.min(dem_utm.y)/1e3, np.max(dem_utm.y)/1e3))
cbar = fig.colorbar(im, ax=ax[0], shrink=0.8, label='Elevation [m]')
ax[0].set_xlabel('Easting [km]')
ax[0].set_ylabel('Northing [km]')

# RGB image
ax[1].imshow(np.dstack([rgb_im['B4'].data, rgb_im['B3'], rgb_im['B2'].data]),
             extent=(min(rgb_im.x)/1e3, max(rgb_im.x)/1e3, min(rgb_im.y)/1e3, max(rgb_im.y)/1e3))
ax[1].set_xlabel('Easting [km]')
ax[1].set_yticklabels([])

# classified image
im = ax[2].imshow(classified_utm.data, cmap=cmap, clim=(0.5,5.5),
                  extent=(np.min(classified_utm.x)/1e3, np.max(classified_utm.x)/1e3, np.min(classified_utm.y)/1e3, np.max(classified_utm.y)/1e3))
cbar = fig.colorbar(im, ax=ax[2], shrink=0.8, ticks=[5, 4, 3, 2])
cbar.ax.set_yticklabels(['Water', 'Rock', 'Ice/firn', 'Snow'])
cbar.ax.set_ylim(1.5,5.5)
cbar.ax.invert_yaxis()
ax[2].set_yticklabels([])
ax[2].set_xlabel('Easting [km]')
ax[2].set_ylabel('')

ax[1].set_xlim(ax[2].get_xlim())
ax[1].set_ylim(ax[2].get_ylim())
for axis in ax[0:3]:
    axis.set_yticks([6538, 6540, 6542, 6544])

# SLA contours
x_mesh, y_mesh = np.meshgrid(np.divide(dem_utm.x.data, 1e3), np.divide(dem_utm.y.data, 1e3))
for axis in ax[0:3]:
    axis.contour(x_mesh, y_mesh, dem_utm.data, levels=[sla], colors='k', linewidth=lw)

# histograms and bounds
bin_edges = np.arange(700, 1481, step=10)
bin_centers = (bin_edges[1:] + bin_edges[0:-1]) / 2
# all elevations
counts, _ = np.histogram(dem_utm.data, bins=bin_edges)
areas = counts * 10**2 / 1e6 # km^2
ax[3].bar(bin_centers, areas, width=bin_edges[1]-bin_edges[0], facecolor='gray', 
          edgecolor='k', alpha=0.5, linewidth=lw-1, label='All elevations')
# snow-covered elevations
dem_snow = xr.where(classified_utm==1, dem_utm, np.nan)
counts, _ = np.histogram(dem_snow.data, bins=bin_edges)
areas = counts * 10**2 / 1e6 # km^2
ax[3].bar(bin_centers, areas, width=bin_edges[1]-bin_edges[0], facecolor=cmap_dict['Snow'], 
          edgecolor='k', alpha=1.0, linewidth=lw-1, label='Snow-covered elevations')
ax[3].axvline(sla, color='k', linewidth=lw, label='Original SLA')
ax[3].axvline(sla_lower_bound, color='k', linestyle='--', linewidth=lw, label='SLA lower bound')
ax[3].axvline(sla_upper_bound, color='k', linestyle=':', linewidth=lw, label='SLA upper bound')
ax[3].text(850, 0.22, "Snow-covered area \nbelow SLA = 0.78 km$^2$", fontsize=9, ha='center',
           bbox=dict(facecolor='w', edgecolor='None', alpha=0.7))
ax[3].text(1320, 0.22, "Snow-free area \nabove SLA = 0.66 km$^2$", fontsize=9, ha='center')
ax[3].set_xlim(np.min(bin_edges)-20, np.max(bin_edges)+20)
ax[3].set_xlabel('Elevation [m]')
ax[3].set_ylabel('Area [km$^2$]')
ax[3].legend(loc='upper left')
ax[3].set_xlim(700, 1400)

# histogram of SLA ranges
ax[4].hist(sla_bounds['SLA_bounds_range'], bins=np.linspace(0, 1000, num=101), 
           facecolor='gray', alpha=0.9, edgecolor='k', linewidth=0.5)
range_median = np.nanmedian(sla_bounds['SLA_bounds_range'])
range_mean = np.nanmean(sla_bounds['SLA_bounds_range'])
range_p25 = np.nanpercentile(sla_bounds['SLA_bounds_range'], 25)
range_p75 = np.nanpercentile(sla_bounds['SLA_bounds_range'], 75)
ax[4].axvline(range_median, color='k', linewidth=lw, 
              label=f"Median = {int(range_median)} m")
ax[4].axvline(range_mean, color='k', linewidth=lw, linestyle='--', 
              label=f"Mean = {int(range_mean)} m")
ax[4].fill_between([range_p25, range_p75], [0,0], [3800,3800], color='k', alpha=0.1, edgecolor='k', linewidth=lw+1,
                   label=f"IQR = {int(range_p25)}–{int(range_p75)} m")
ax[4].legend(loc='center right')
ax[4].set_xlabel('Range of all SLA bounds [m]')
ax[4].set_ylabel('Counts')
ax[4].set_xlim(0, 500)
ax[4].set_ylim(0, 3800)
# add panel labels
labels = ['a', 'b', 'c', 'd', 'e']
for i, axis in enumerate(ax):
    if i < 3:
        xscale=0.8
        yscale=0.85
    else:
        xscale=0.95
        yscale=0.9
    axis.text(xscale, yscale, labels[i], transform=axis.transAxes, fontsize=fontsize+4, fontweight='bold',
              bbox=dict(facecolor='w', edgecolor='w'))

# fig.tight_layout()

ax[1].set_position([0.44, ax[0].get_position().y0, ax[0].get_position().width, ax[0].get_position().height])

plt.show()

# Save figure to file
fig_fn = os.path.join(figures_out_path, 'figS2_SLA_uncertainties.png')
save_figure(fig, fig_fn)


## Figure S3. Median snow cover minima sampling

# Set up figure
gs = matplotlib.gridspec.GridSpec(2,2, height_ratios=[1.5, 1])
fig = plt.figure(figsize=(7,8))
ax = [fig.add_subplot(gs[0,:]), 
      fig.add_subplot(gs[1,0]), fig.add_subplot(gs[1,1])]

# Plot an example Monte Carlo simulation at South Cascade Glacier
rgi_id = 'RGI60-02.18778'
scs_fn = os.path.join(scm_path, 'study-sites', rgi_id, f"{rgi_id}_classifications.zarr")
scs = f.load_snow_cover_stats(scs_fn)
scs = scs.assign_coords({'WOY': scs['time'].dt.isocalendar().week,
                         'Year': scs['time'].dt.isocalendar().year})
scs['snow_area_km2'] = scs['snow_area'] / 1e6

# Plot
sns.scatterplot(scs, x='WOY', y='AAR', hue='Year', palette='viridis', s=10, legend=True, ax=ax[0])
# Number of samples per simulation
sample_fraction = 0.8
nsamp = int(len(scs) * sample_fraction)
nMC = 100
# Monte Carlo simulations
results = pd.DataFrame()
for i in range(nMC):
    sampled_indices = np.random.choice(scs.time.data, size=nsamp, replace=False)
    scs_MC = scs.isel(time=i)
    # Calculate weekly medians for AAR, SCA, and SLA
    weekly_medians = scs[['AAR', 'snow_area_km2', 'SLA']].groupby('WOY').median().to_dataframe()
    weekly_medians['MC_run'] = i    
    # Plot
    ax[0].plot(weekly_medians.index, weekly_medians['AAR'], '-', color='gray', linewidth=1)
    results = pd.concat([results, weekly_medians])


# Dummy line for legend
ax[0].plot([0,0], [0,0], '-', color='gray', linewidth=2, label='MC simulation')
# Estimate median AAR and snow minimum timing
medians = results.groupby('WOY')['AAR'].median()
value = medians.loc[medians==min(medians)]
ax[0].scatter(value.index[0], medians.min(), marker='x', s=50, color='k', linewidth=2, label='Median AAR', zorder=5)
handles, labels = ax[0].get_legend_handles_labels()
ax[0].set_ylabel('Transient accumulation area ratio')
ax[0].set_xlabel('Week of year')
ax[0].set_xlim(16,45)
ax[0].set_ylim(-0.1,1.1)
# add legend
ax[0].legend().remove()
handles, labels = ax[0].get_legend_handles_labels()
ax[0].legend(handles, labels, loc='lower left', markerscale=2)

# Plot AAR and snow minima timing distributions for all sites
median_scs_fn = os.path.join(out_path, 'median_snow_cover_stats_MC.nc')
median_scs = xr.open_dataset(median_scs_fn)
aar_woy_mads = pd.DataFrame()
# iterate over site names in median snow cover stats dataframe
for rgi_id in tqdm(median_scs.RGIId.values):
    # Calculate AAR median and MAD across MC simulations
    aar_median = float(median_scs.sel(RGIId=rgi_id).AAR.min(dim='WOY').median().values)
    aar_mad = median_abs_deviation(median_scs.sel(RGIId=rgi_id).AAR.min(dim='WOY').values)
    Imins = median_scs.sel(RGIId=rgi_id).AAR.argmin(dim='WOY').values
    woy_mad = median_abs_deviation(median_scs.WOY.values[Imins])
    # Add to dataFrame
    df = pd.DataFrame({'RGIId': [rgi_id],
                       'AAR_MAD': [aar_mad],
                       'WOY_MAD': [woy_mad]})
    aar_woy_mads = pd.concat([aar_woy_mads, df], axis=0)


# Plot
ax[1].hist(aar_woy_mads['AAR_MAD'], bins=50, color='gray', edgecolor='k', linewidth=0.5)
ax[1].text(0.95, 0.7, f"Median = {np.round(np.nanmedian(aar_woy_mads['AAR_MAD']),2)}", ha='right', transform=ax[1].transAxes)
ax[1].text(0.95, 0.63, f"Mean = {np.round(np.nanmean(aar_woy_mads['AAR_MAD']),2)}", ha='right', transform=ax[1].transAxes)
ax[1].set_xlabel('AAR MAD')
ax[1].set_ylabel('Counts')
ax[2].hist(aar_woy_mads['WOY_MAD'], bins=50, color='gray', edgecolor='k', linewidth=0.5)
ax[2].text(0.95, 0.7, f"Median = {np.round(np.nanmedian(aar_woy_mads['WOY_MAD']),2)}", ha='right', transform=ax[2].transAxes)
ax[2].text(0.95, 0.63, f"Mean = {np.round(np.nanmean(aar_woy_mads['WOY_MAD']),2)}", ha='right', transform=ax[2].transAxes)
ax[2].set_xlabel('Snow minimum timing MAD [weeks]')
ax[2].set_xlim(-0.1,3.5)

# add panel labels
labels = ['a', 'b', 'c']
for i, axis in enumerate(ax):
    axis.text(0.9, 0.92, labels[i], transform=axis.transAxes, fontsize=fontsize+2, fontweight='bold')


fig.tight_layout()
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'figS3_AAR_MC_simulations.png')
save_figure(fig, fig_fn)

## Figure S4. Climate clustering

# Load climate clusters
clusters_fn = os.path.join(out_path, 'climate_clusters.csv')
clusters = pd.read_csv(clusters_fn)
print('Clusters loaded')

# Load feature scaler
scaler_fn = os.path.join(code_path, 'inputs-outputs', 'ERA5_feature_scaler.joblib')
scaler = joblib.load(scaler_fn)

# Apply the feature scaler for normalized values
clusters[['mean_annual_precip_cumsum_norm', 'mean_annual_temp_range_norm']] = scaler.transform(clusters[['mean_annual_precip_cumsum', 'mean_annual_temp_range']])
print('Added standardized feature columns')

# Load AOIs
aois = gpd.read_file(aois_fn)
# Add climate clusters column
aois = pd.merge(aois, clusters[['RGIId', 'clustName']], on='RGIId')
print('AOIs loaded')

# Load RGI O2 Regions
rgi_O2_fn = os.path.join(scm_path, '..', 'GIS_data', 'RGI', 'RGIv7_02Regions', 
                                'RGI2000-v7.0-o2regions-Alaska-westernCanadaUS_clipped_to_country_outlines.shp')
rgi_O2 = gpd.read_file(rgi_O2_fn)
# remove Brooks Range
rgi_O2 = rgi_O2.loc[rgi_O2['o2region']!='01-01']
# add subregion name and color column
rgi_O2[['Subregion', 'color']] = '', ''
for i, o1o2 in enumerate(rgi_O2['o2region'].drop_duplicates().values):
    o1 = int(o1o2[0:2])
    o2 = int(o1o2[3:])
    subregion_name, color = f.determine_subregion_name_color(o1, o2)
    rgi_O2.loc[rgi_O2['o2region']==o1o2, 'Subregion'] = subregion_name
print('RGI O2 regions loaded')

# Load GTOPO30
gtopo_fn = '/Users/raineyaberle/Research/PhD/GIS_data/GTOPO30_clip.tif'
gtopo = rxr.open_rasterio(gtopo_fn)
gtopo = xr.where(gtopo==-32768, np.nan, gtopo)
print('GTOPO30 loaded')

# Get limits for scaled data
fig, ax = plt.subplots()
sns.scatterplot(data=clusters, x='mean_annual_temp_range_norm', y='mean_annual_precip_cumsum_norm')
xmin, xmax = ax.get_xlim()
ymin, ymax = ax.get_ylim()
plt.close()

# Set up figure
ontsize = 12
plt.rcParams.update({'font.size': fontsize, 'font.sans-serif': 'Arial'})
gs = matplotlib.gridspec.GridSpec(2,2, height_ratios=[1,1.5])
fig = plt.figure(figsize=(10,10))
ax = [fig.add_subplot(gs[0,0]), fig.add_subplot(gs[0,1]), fig.add_subplot(gs[1,:])]

### a) Feature values and cluster assignments
# Input features and scaled features
scat = sns.scatterplot(data=clusters, x='mean_annual_temp_range', y='mean_annual_precip_cumsum',
                       edgecolor='k', linewidth=0.5, s=20, 
                       hue='clustName', hue_order=cluster_order, palette=cluster_cmap_dict, legend=False, ax=ax[0])
ax[0].grid()
ax[0].set_xlabel('Air temperature range [$^{\circ}$C]')
ax[0].set_ylabel('Precipitation sum [m]')
# add secondary axes for scaled features
ax_top = ax[0].twiny()
ax_right = ax[0].twinx()
ax_top.set_xlim(xmin, xmax)
ax_right.set_ylim(ymin, ymax)
ax_right.set_ylabel('Standardized air temperature range [unitless]', color='grey')
ax_top.set_xlabel('Standardized precipitation sum [unitless]', color='grey')
ax_right.spines['top'].set_color('grey')
ax_right.spines['right'].set_color('grey')
ax_top.set_xticklabels(ax_top.get_xticklabels(), color='grey')
ax_right.set_yticklabels(ax_right.get_yticklabels(), color='grey')
ax_top.tick_params(axis='x', colors='grey')
ax_right.tick_params(axis='y', colors='grey')

# Silhouette coefficient and intertia
K = np.arange(2, 11)
# copied outputs from notebook: 2_develop_climate_clusters.ipynb
sil_coefs = [0.426, 0.419, 0.442, 0.460, 0.446, 0.415, 0.430, 0.439, 0.422]
inertias = [198, 132, 91, 68, 56, 48, 41, 32, 30]
inertia_color = '#d95f02'
sil_color = '#7570b3'
# silhouette coefficient
ax[1].plot(K, sil_coefs, '.-', color=sil_color)
Ibest = np.argwhere(sil_coefs==np.nanmax(sil_coefs))[0][0]
ax[1].plot(K[Ibest], sil_coefs[Ibest], '*', color=sil_color, markersize=15)
ax[1].set_xlabel('Number of clusters')
ax[1].set_ylabel('Silhouette score', color=sil_color)
ax[1].grid()
ax[1].tick_params(axis='y', color=sil_color)
ax[1].set_yticklabels(ax[1].get_yticklabels(), color=sil_color)
# inertia
ax2 = ax[1].twinx()
ax2.plot(K, inertias, '.-', color=inertia_color)
Ibest = 3
ax2.plot(K[Ibest], inertias[Ibest], '*', color=inertia_color, markersize=15)
ax2.spines['right'].set_color(inertia_color)
ax2.spines['left'].set_color(sil_color)
ax2.set_ylabel('Inertia', color=inertia_color)
ax2.tick_params(axis='y', color=inertia_color)
ax2.set_yticklabels(ax2.get_yticklabels(), color=inertia_color)
ax2.text(4.8, 68, 'elbow', color=inertia_color, rotation=-35, fontsize=14)

### c) Map view
# GTOPO hillshade
ls = matplotlib.colors.LightSource(azdeg=90, altdeg=45)
ax[2].imshow(ls.hillshade(gtopo.data[0], vert_exag=0.002), cmap='gray', alpha=0.5,
             extent=(np.min(gtopo.x.data), np.max(gtopo.x.data), 
                     np.min(gtopo.y.data), np.max(gtopo.y.data)))
# RGI O2 region outlines
color = '#525252'
rgi_O2.plot(ax=ax[2], alpha=1.0, facecolor='None', edgecolor=color, linewidth=1)
ax[2].set_yticks(np.linspace(45, 65, num=6))
ax[2].set_xlim(-167, -112)
ax[2].set_ylim(46, 66)
ax[2].set_xlabel('Longitude ($^{\circ}$E)')
ax[2].set_ylabel('Latitude ($^{\circ}$N)')
ax[2].set_aspect(2.2)
# Site locations
sns.scatterplot(data=aois, x='CenLon', y='CenLat', edgecolor='k', linewidth=0.5, s=20,
                hue='clustName', palette=cluster_cmap_dict, hue_order=cluster_order, alpha=1, ax=ax[2])
handles, labels = ax[2].get_legend_handles_labels()
ax[2].legend().remove()
ax[2].legend(handles, labels, loc='lower left', markerscale=2, bbox_to_anchor=[0.08, 0.08, 0.2, 0.2])
# Add region labels and arrows
fontweight = 'bold'
background_color = [1, 1, 1, 0.5]
ax[2].text(-160.5, 56.5, "Aleutians", ha='center', color=color, rotation=35, fontsize=fontsize-3, fontweight=fontweight)
ax[2].text(-157, 62.7, "Alaska Range", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].text(-147.7, 57.7, "W. Chugach \nMtns.", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].arrow(-147.6, 58.9, 0, 0.6, color=color, linewidth=2, head_width=0.34, head_length=0.2)
ax[2].text(-141.5, 57.7, "St. Elias \nMtns.", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].arrow(-141.5, 58.9, 0, 0.6, color=color, linewidth=2, head_width=0.34, head_length=0.2)
ax[2].text(-139.8, 56, "N. Coast \nRanges", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].arrow(-137.1, 56.5, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[2].text(-133.7, 51.3, "N. Cascades", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].arrow(-129.9, 51.4, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[2].text(-130, 47.4, "S. Cascades", ha='center', color=color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].arrow(-126.3, 47.5, 1.2, 0, color=color, linewidth=2, head_width=0.2, head_length=0.3)
ax[2].text(-129.5, 64.5, "N. Rockies", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].text(-119, 54.5, "C. Rockies", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)
ax[2].text(-115, 47, "S. Rockies", ha='center', color=color, backgroundcolor=background_color, rotation=0, fontsize=fontsize-3, fontweight=fontweight)

# Add panel labels
labels = ['a', 'b', 'c']
for i, axis in enumerate(ax):
    axis.text(axis.get_xlim()[0] + (axis.get_xlim()[1] - axis.get_xlim()[0])*0.9,
              axis.get_ylim()[0] + (axis.get_ylim()[1] - axis.get_ylim()[0])*0.9,
              labels[i], fontweight='bold', fontsize=fontsize+4, ha='center')

fig.tight_layout()
plt.show()

# Save figure to file
fig_fn = os.path.join(figures_out_path, 'figS4_kmeans.png')
save_figure(fig, fig_fn)

## Figure S5. Constraining mass balance model parameters

# Load glacier IDs from model runs
pygem_new_path = os.path.join(scm_path, 'Brandon_new_PyGEM_runs')
rgi_ids = [x for x in sorted(os.listdir(pygem_new_path)) if os.path.isdir(os.path.join(pygem_new_path, x))]
names = ['Gulkana', 'Wolverine', 'Lemon Creek', 'Sperry', 'South Cascade']
print('RGI IDs for glaciers with PyGEM runs:', rgi_ids)
# Define path to original model runs
model_path = os.path.join(scm_path, 'Rounce_et_al_2023')

# Set up figure
fig, ax = plt.subplots(len(rgi_ids), 2, figsize=(8, 12), gridspec_kw={'width_ratios': [1.5,1]})
rmse_cmap = 'Blues'
observed_color = 'k'
original_color = '#fec44f'
adjusted_color = '#993404'

# Iterate over RGI IDs
for i, rgi_id in enumerate(rgi_ids):
    name = names[i]
    print(name, rgi_id)

    # Load observed snow cover data
    scs_fn = os.path.join(scm_path, 'study-sites', f"RGI60-0{rgi_id}", f"RGI60-0{rgi_id}_snow_cover_stats_adjusted.csv")
    scs = pd.read_csv(scs_fn)
    scs['datetime'] = pd.to_datetime(scs['datetime'])
    
    # Load PyGEM runs
    runs_fn = os.path.join(scm_path, 'analysis', f"PyGEM_comparison_RGI60-0{rgi_id}.nc")
    runs = xr.open_dataset(runs_fn)

    # Load original model parameters
    modelprms_fn = os.path.join(model_path, '..', 'Rounce_et_al_2023', 'modelprms', f"{rgi_id}-modelprms_dict.pkl")
    modelprms = pd.read_pickle(modelprms_fn)
    # get the original values
    original_ddfsnow = np.median(modelprms['MCMC']['ddfsnow']['chain_0'])
    original_tbias = np.median(modelprms['MCMC']['tbias']['chain_0'])
    original_kp = np.median(modelprms['MCMC']['kp']['chain_0'])

    # Calculate RMSE for each run's snowline altitudes
    diff = runs['mod-obs_SLA']
    rmse_by_run = np.sqrt((diff**2).mean(dim='time'))
    runs['rmse'] = rmse_by_run

    # identify parameter combinations with the lowest RMSE
    df_plot = runs[['tbias', 'ddfsnow', 'kp', 'rmse', 'mod-obs_SLA']].to_dataframe().reset_index()
    df_plot = df_plot.dropna(subset=['rmse'])
    df_plot_best = df_plot.loc[df_plot['rmse'].idxmin()]

    # subset the model dataset for the original vs. best runs
    # original
    squared_diffs = sum((runs[var] - np.median(modelprms['MCMC'][var]['chain_0']))**2 for var in ['tbias', 'ddfsnow', 'kp'])
    best_run_idx = squared_diffs.argmin(dim="run")
    combined_ds_original = runs.sel(run=best_run_idx, glac=0)
    # best
    combined_ds_best = runs.sel(run=int(df_plot_best['run']), glac=0)

    # adjust ddf units for plotting
    combined_ds_original['ddfsnow'] = combined_ds_original['ddfsnow'] * 1e3
    combined_ds_best['ddfsnow'] = combined_ds_best['ddfsnow']
    df_plot['ddfsnow'] = df_plot['ddfsnow'] * 1e3
    df_plot_best['ddfsnow'] = df_plot_best['ddfsnow'] * 1e3
    original_ddfsnow = original_ddfsnow * 1e3

    # add plotting columns
    df_plot['size'] = df_plot['kp']*10
    # df_plot.sort_values(by='rmse', inplace=True, ascending=False) # plot lowest RMSE on top    

    ### Plot RMSE as a function of tbias, ddfsnow, and kp
    im = ax[i,0].scatter(df_plot['tbias'], df_plot['ddfsnow'], s=df_plot['size'], c=df_plot['rmse'], cmap=rmse_cmap, 
                         marker='o', edgecolor='k', linewidth=0.25)
    # add colorbar
    axin = inset_axes(ax[i,0], width="20%", height="4%", loc="lower left")
    cbar = plt.colorbar(im, cax=axin, orientation="horizontal")
    axin.xaxis.set_ticks_position('top')
    cbar.ax.set_title('RMSE [m]', fontsize=9)
    cbar.ax.tick_params(labelsize=8)
    # original parameter combinations
    ax[i,0].plot(original_tbias, original_ddfsnow, 's', markersize=10, markeredgecolor=original_color, 
                 markerfacecolor='None', markeredgewidth=2, label='Original')
    # best parameter combination
    ax[i,0].plot(df_plot_best['tbias'], df_plot_best['ddfsnow'], 
                 '*', markersize=15, markeredgecolor=adjusted_color, markerfacecolor='None', markeredgewidth=2, label='Adjusted')
    ax[i,0].grid(True)
    ax[i,0].set_ylabel('DDF$_{snow}$ [mm $^{\circ}$C$^{-1}$ d$^{-1}$]')
    ax[i,0].set_title(f'{name} Glacier')   
    ax[i,0].set_ylim(0, 10)
    # ax[i,0].set_xlim(-7,4)

    ### Modeled and observed snowline time series by WOY
    # original
    mod_og_gb = combined_ds_original['glac_snowline_monthly'].groupby(combined_ds_original.time.dt.isocalendar().week).quantile([0.25, 0.5, 0.75])
    mod_og_p25, mod_og_p50, mod_og_p75 = mod_og_gb.sel(quantile=0.25), mod_og_gb.sel(quantile=0.50), mod_og_gb.sel(quantile=0.75)
    ax[i,1].fill_between(mod_og_p25.week, mod_og_p25, mod_og_p75, facecolor=original_color, edgecolor='None', alpha=0.2)
    ax[i,1].plot(mod_og_p50.week, mod_og_p50.values, '-', color=original_color, label='Original')
    # adjusted
    mod_best_gb = combined_ds_best['glac_snowline_monthly'].groupby(combined_ds_best.time.dt.isocalendar().week).quantile([0.25, 0.5, 0.75])
    mod_best_p25, mod_best_p50, mod_best_p75 = mod_best_gb.sel(quantile=0.25), mod_best_gb.sel(quantile=0.50), mod_best_gb.sel(quantile=0.75)
    ax[i,1].fill_between(mod_best_p25.week, mod_best_p25, mod_best_p75, facecolor=adjusted_color, edgecolor='None', alpha=0.2)
    ax[i,1].plot(mod_best_p50.week, mod_best_p50.values, '--', color=adjusted_color, label='Adjusted')
    # observed
    scs_p25 = scs.groupby(scs['datetime'].dt.isocalendar().week)['SLA_m'].quantile(0.25).reset_index()
    scs_p50 = scs.groupby(scs['datetime'].dt.isocalendar().week)['SLA_m'].quantile(0.50).reset_index()
    scs_p75 = scs.groupby(scs['datetime'].dt.isocalendar().week)['SLA_m'].quantile(0.75).reset_index()
    ax[i,1].fill_between(scs_p25['week'].astype(float), scs_p25['SLA_m'], scs_p75['SLA_m'], color=observed_color, alpha=0.2)
    ax[i,1].plot(scs_p50['week'].astype(float), scs_p50['SLA_m'], ':', color=observed_color, label='Observed')
    ax[i,1].set_xlim(17, 41)
    ax[i,1].set_ylabel('Elevation [m]')

    # add x-labels and legends on last row
    if i==len(rgi_ids)-1:
        ax[i,0].set_xlabel('Temperature bias [$^{\circ}$C]')
        ax[i,1].set_xlabel('Week of year')
        # custom frame for first two legends
        patch = matplotlib.patches.FancyBboxPatch((-4.5, -10.4), 5.5, 6.1, facecolor='None', edgecolor='gray', 
                                                  linewidth=0.5, clip_on=False, boxstyle=matplotlib.patches.BoxStyle("round", pad=0.2))
        ax[4,0].add_patch(patch)
        # first legend
        first_legend = ax[i,0].legend(bbox_to_anchor=[0.1, -0.66, 0.2, 0.2], frameon=False)
        ax[i,0].add_artist(first_legend)
        # second legend for kp (using dummy points)
        xmin, xmax = ax[i,0].get_xlim()
        for s in [1,2,3]:
            ax[i,0].plot(-100, -100, 'ok', markersize=s*3, label=str(s))
        ax[i,0].set_xlim(xmin, xmax)
        handles, labels = ax[i,0].get_legend_handles_labels()
        ax[i,0].legend(handles[2:], labels[2:], loc='upper center', bbox_to_anchor=[0.6, -0.53, 0.2, 0.2], 
                       frameon=False, title='Precipitation factor')
        # third legend
        ax[i,1].legend(loc='upper center', bbox_to_anchor=[0.4, -0.5, 0.2, 0.2])
    else:
        ax[i,0].set_xlabel('')
        ax[i,1].set_xlabel('')

# add panel labels
import string
labels = list(string.ascii_lowercase)
for i, axis in enumerate(ax.ravel()):
    if i % 2 == 0:
        xpos = 0.95
    else:
        xpos = 0.05
    axis.text(xpos, 0.8, labels[i], transform=axis.transAxes, fontweight='bold', fontsize=14)

fig.tight_layout()
plt.show()

# Save to file
fig_fn = os.path.join(figures_out_path, 'figS5_PyGEM_constraints.png')
save_figure(fig, fig_fn)

## Table S1. Study sites with subregion and climate cluster

aois = gpd.read_file(aois_fn)
aois.rename(columns={'Subregion': 'Subregion name'}, inplace=True)
aois[['O1Region', 'O2Region']] = aois[['O1Region', 'O2Region']].astype(int)

# Add climate cluster
clusters = pd.read_csv(clusters_fn)
aois['Climate cluster'] = ''
for rgi_id in aois['RGIId'].drop_duplicates().values:
    aois.loc[aois['RGIId']==rgi_id, 'Climate cluster'] = clusters.loc[clusters['RGIId']==rgi_id, 'clustName']
aois.sort_values(by=['O1Region', 'O2Region'], inplace=True)

# Format as LaTeX table
columns = ['RGIId', 'O1Region', 'O2Region', 'Subregion name', 'Climate cluster']
aois = aois[columns]

# Save as Excel sheet
aois_xl_fn = os.path.join(out_path, 'TableS1_study_sites.xlsx')
aois.to_excel(aois_xl_fn, index=False)
print('Table saved as Excel spreadsheet:', aois_xl_fn)
aois


## Glacier area changes

# Calculate minimum annual areas distribution
# Load study sites
aois = gpd.read_file(aois_fn)
# Load climate clusters
clusters = pd.read_csv(clusters_fn)

# Initialize dataframe to store area distribution results
dfs_list = []

# Iterate over study sites
for rgi_id in tqdm(aois['RGIId'].unique()):
    # load snow cover stats
    scs_fn = os.path.join(scm_path, 'study-sites', rgi_id, f"{rgi_id}_snow_cover_stats_adjusted2.csv")
    scs = pd.read_csv(scs_fn)
    scs['datetime'] = pd.to_datetime(scs['datetime'])
    scs = scs.loc[scs['datetime'].dt.year >= 2016]
    scs['glacier_area_km2'] = scs['glacier_area_m2'] / 1e6    

    # subset to later in the season
    scs = scs.loc[(scs['datetime'].dt.month >= 8) & (scs['datetime'].dt.month <= 10)]

    # Calculate 25th, 50th, and 75th percentiles of minimum annual area
    area_q25, area_q50, area_q75 = np.nanpercentile(scs.groupby(scs['datetime'].dt.year)['glacier_area_km2'].median(), 
                                                    [25, 50, 75])
    
    # Calculate as fraction of RGI area
    rgi_area = aois.loc[aois['RGIId']==rgi_id, 'Area'].values[0]
    area_q25_frac, area_q50_frac, area_q75_frac = np.array([area_q25, area_q50, area_q75]) / rgi_area

    # Store in dataframe
    df = pd.DataFrame({'RGIId': [rgi_id],
                       'clustName': [clusters.loc[clusters['RGIId']==rgi_id, 'clustName'].values[0]],
                       'Subregion': [aois.loc[aois['RGIId']==rgi_id, 'Subregion'].values[0]],
                       'RGI_area_km2': [rgi_area],
                       'glacier_area_km2_P25': [area_q25],
                       'glacier_area_km2_P50': [area_q50],
                       'glacier_area_km2_P75': [area_q75],
                       'glacier_area_fraction_P25': [area_q25_frac],
                       'glacier_area_fraction_P50': [area_q50_frac],
                       'glacier_area_fraction_P75': [area_q75_frac],
                       })
    
    # Add to dataframe list
    dfs_list.append(df)

# Concatenate all dataframes
areas_df = pd.concat(dfs_list)
areas_df.reset_index(drop=True, inplace=True)
areas_df

# Plot distribution for each site
fig, ax = plt.subplots(1, 1, figsize=(16,5))
for rgi_id in areas_df['RGIId'].unique():
    id_display = rgi_id.replace('RGI60-0','')
    area_df = areas_df.loc[areas_df['RGIId']==rgi_id]
    ax.plot([id_display, id_display], 
            [area_df['glacier_area_fraction_P25'], area_df['glacier_area_fraction_P75']],
            '-', color='gray')
    ax.plot(id_display, area_df['glacier_area_fraction_P50'], '.k')

ax.set_xticklabels(ax.get_xticklabels(), rotation=90, fontsize=5)
ax.set_xlim(-2, 202)
ax.set_ylim(0, 0.9)
ax.set_ylabel('Fraction of RGI v. 6 glacier area')
plt.show()

# plot boxplot of IQRs by climate cluster
sns.boxplot(areas_df, x='clustName', y='glacier_area_fraction_P50', hue='clustName', palette=cluster_cmap_dict, showfliers=False)

# plot boxplot of IQRs by subregion
plt.rcParams.update({'font.size':12, 'font.sans-serif': 'Arial'})
fig, ax = plt.subplots(1, 1, figsize=(6,6))
sns.boxplot(areas_df, y='Subregion', x='glacier_area_fraction_P50', 
            hue='Subregion', palette='mako', medianprops=dict(color="w", linewidth=2), showfliers=False, ax=ax)
ax.set_xlabel('Median glacier area / RGI v. 6 area')
ax.set_ylabel('RGI O2 region')
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'glacier_area_change_2016–2023.png')
save_figure(fig, fig_fn)

# Print statistics by subregion
print(areas_df['glacier_area_fraction_P50'].describe())
areas_df.groupby('Subregion')['glacier_area_fraction_P50'].describe()

## Number of observations at each site vs. region

aois = gpd.read_file(aois_fn)
clusters = pd.read_csv(clusters_fn)

df_list = []
for i in range(len(aois)):
    aoi = aois.iloc[i]
    rgi_id = aoi['RGIId']
    cluster = clusters.loc[clusters['RGIId']==rgi_id, 'clustName'].values[0]
    classified_fns = sorted(glob.glob(os.path.join(scm_path, 'study-sites', rgi_id, 'classified', '*classified.nc')))
    df_list.append(pd.DataFrame({'RGIId': rgi_id,
                                 'Subregion': [aoi['Subregion']],
                                 'clustName': [cluster],
                                 'nobs': [len(classified_fns)]}))
    
nobs_df = pd.concat(df_list)
nobs_df.reset_index(drop=True, inplace=True)
nobs_df

sns.boxplot(nobs_df, y='clustName', x='nobs', hue='clustName', 
            palette=cluster_cmap_dict, hue_order=cluster_order)

plt.rcParams.update({'font.sans-serif': 'Arial', 'font.size': 12})
fig, ax = plt.subplots()
sns.boxplot(nobs_df, y='Subregion', x='nobs', hue='Subregion', 
            palette='mako', hue_order=subregion_order, ax=ax)
ax.set_ylabel('RGI O2 region')
ax.set_xlabel('Number of observations')
plt.show()

nobs_df.groupby('Subregion')['nobs'].describe().sort_values(by='50%')

fig_fn = os.path.join(figures_out_path, 'number_of_observations_by_region.png')
save_figure(fig, fig_fn)

## README figures

### Median weekly trends

# Set up figure
plt.rcParams.update({'font.size':14, 'font.sans-serif': "Arial"})
fig, ax = plt.subplots(1, 3, figsize=(15,5))

# Load snow cover stats
rgi_id = 'RGI60-02.18778'
scs_fn = os.path.join(scm_path, 'study-sites', rgi_id, f"{rgi_id}_snow_cover_stats.csv")
scs = pd.read_csv(scs_fn)
scs['datetime'] = pd.to_datetime(scs['datetime'], errors='coerce')
scs.dropna(subset=['datetime'], inplace=True)
scs['WOY'] = scs['datetime'].dt.isocalendar().week
scs['Year'] = scs['datetime'].dt.isocalendar().year
scs['Year'] = pd.Categorical(scs['Year'])
scs['SCA_km2'] = scs['snow_area'] / 1e6
# Plot
sns.scatterplot(scs, x='WOY', y='AAR', hue='Year', palette='viridis', size=0.5, legend=False, ax=ax[0])
sns.scatterplot(scs, x='WOY', y='SCA_km2', hue='Year', palette='viridis', size=0.5, legend=False, ax=ax[1])
sns.scatterplot(scs, x='WOY', y='ELA_from_AAR', hue='Year', palette='viridis', size=0.5, legend=False, ax=ax[2])


# Number of samples per simulation
sample_fraction = 0.8
nsamp = int(len(scs) * sample_fraction)
nMC = 100

# Monte Carlo simulations
results = pd.DataFrame()
for i in range(nMC):
    sampled_indices = np.random.choice(scs.index, size=nsamp, replace=False)
    scs_MC = scs.loc[sampled_indices].sort_values(by='datetime')

    # Calculate weekly medians for AAR, SCA, and ELA
    weekly_medians = scs_MC.groupby('WOY')[['AAR', 'SCA_km2', 'ELA_from_AAR']].median()
    weekly_medians['MC_run'] = i    
    # Plot
    if i==0:
        label = 'MC simulations'
    else:
        label = '_nolegend'
    ax[0].plot(weekly_medians.index, weekly_medians['AAR'], '-', color='gray', linewidth=0.1, label=label)
    ax[1].plot(weekly_medians.index, weekly_medians['SCA_km2'], '-', color='gray', linewidth=0.1)
    ax[2].plot(weekly_medians.index, weekly_medians['ELA_from_AAR'], '-', color='gray', linewidth=0.1)
    results = pd.concat([results, weekly_medians])
    
# Estimate median AAR and snow minimum timing
for i, column in enumerate(['AAR', 'SCA_km2', 'ELA_from_AAR']):
    medians = results.groupby('WOY')[column].median()
    if column!='ELA_from_AAR':
        value = medians.loc[medians==min(medians)]
    else:
        value = medians.loc[medians==max(medians)]
    ax[i].axvline(value.index[0], color='k', label='Minimum snow cover median')
handles, labels = ax[0].get_legend_handles_labels()
ax[0].set_ylabel('Transient accumulation area ratio')
ax[1].set_ylabel('Snow covered area [km$^2$]')
ax[2].set_ylabel('Snowline altitude [m]')
for axis in ax:
    axis.set_xlabel('Week of year')
fig.tight_layout()
plt.show()

# Save figure
# fig_fn = os.path.join(figures_out_path, 'weekly_median_trends_example.png')
# save_figure(fig, fig_fn)

## AGU24 figures

### Abstract figure

cmap = sns.color_palette('mako', n_colors=len(subregion_order)+2)

# -----Load median AARs for all sites
min_snow_cover_stats_fn = os.path.join(scm_path, 'results', 'min_snow_cover_stats.csv')
min_snow_cover_stats = pd.read_csv(min_snow_cover_stats_fn)
# Sort subregions
min_snow_cover_stats['order'] = ''
for i, subregion in enumerate(subregion_order):
    min_snow_cover_stats.loc[min_snow_cover_stats['Subregion']==subregion, 'order'] = i
    min_snow_cover_stats.loc[min_snow_cover_stats['Subregion']==subregion, 'color'] = matplotlib.colors.to_hex(cmap[i])
min_snow_cover_stats = min_snow_cover_stats.sort_values(by='order')
print('Median AARs loaded from file')

# # -----Load RGI O2 Regions
# rgi_O2_fn = os.path.join(scm_path, '..', 'GIS_data', 'RGI', 'RGIv7_02Regions', 
#                                 'RGI2000-v7.0-o2regions-Alaska-westernCanadaUS_clipped_to_country_outlines.shp')
# rgi_O2 = gpd.read_file(rgi_O2_fn)
# # remove Brooks Range
# rgi_O2 = rgi_O2.loc[rgi_O2['o2region']!='01-01']
# # add subregion name and color column
# rgi_O2[['Subregion', 'color']] = '', ''
# for i, o1o2 in enumerate(rgi_O2['o2region'].drop_duplicates().values):
#     o1 = int(o1o2[0:2])
#     o2 = int(o1o2[3:])
#     subregion_name, color = f.determine_subregion_name_color(o1, o2)
#     rgi_O2.loc[rgi_O2['o2region']==o1o2, 'Subregion'] = subregion_name
#     rgi_O2.loc[rgi_O2['o2region']==o1o2, 'color'] = dict(min_snow_cover_stats[['Subregion', 'color']].drop_duplicates().values)[subregion]
# print('RGI O2 regions loaded from file')

# # -----Load GTOPO30
# gtopo_fn = '/Users/raineyaberle/Research/PhD/GIS_data/GTOPO30_clip.tif'
# gtopo = rxr.open_rasterio(gtopo_fn)
# gtopo = xr.where(gtopo==-32768, np.nan, gtopo)
# print('GTOPO30 loaded from file')

# # -----Load classified image
# site_name = 'RGI60-01.00037'
# im_classified_fn = f'/Volumes/LaCie/raineyaberle/Research/PhD/snow_cover_mapping/study-sites/{site_name}/imagery/classified/20230802T152742_RGI60-01.00037_Sentinel-2_SR_classified.nc'
# im_classified = xr.open_dataset(im_classified_fn)
# print('Classified image loaded')

# # -----Load classified images colormap
# import json
# datasets_dict_fn = '/Users/raineyaberle/Research/PhD/snow_cover_mapping/snow-cover-mapping/inputs-outputs/datasets_characteristics.json'
# datasets_dict = json.load(open(datasets_dict_fn))
# cmap_classified = matplotlib.colors.ListedColormap(datasets_dict['classified_image']['class_colors'].values())

# # -----Load Sentinel-2 image from GEE
# import math
# import wxee as wx
# import geedim as gd
# import ee
# ee.Initialize()
# def convert_wgs_to_utm(lon: float, lat: float):
#     utm_band = str((math.floor((lon + 180) / 6) % 60) + 1)
#     if len(utm_band) == 1:
#         utm_band = '0' + utm_band
#     if lat >= 0:
#         epsg_code = '326' + utm_band
#         return epsg_code
#     epsg_code = '327' + utm_band
#     return epsg_code
# # Load AOI
# aoi_fn = glob.glob(f'/Volumes/LaCie/raineyaberle/Research/PhD/snow_cover_mapping/study-sites/{site_name}/AOIs/*.shp')[0]
# aoi = gpd.read_file(aoi_fn)
# aoi_bounds = aoi.geometry[0].bounds
# region = ee.Geometry.Polygon([[aoi_bounds[0], aoi_bounds[1]], 
#                               [aoi_bounds[2], aoi_bounds[1]],
#                               [aoi_bounds[2], aoi_bounds[3]],
#                               [aoi_bounds[0], aoi_bounds[3]],
#                               [aoi_bounds[0], aoi_bounds[1]]])
# # Load image collection
# im_col = gd.MaskedCollection.from_name('COPERNICUS/S2_SR_HARMONIZED').search(start_date='2023-08-01',
#                                                                              end_date='2023-08-03',
#                                                                              region=region,
#                                                                              mask=True)
# im_ee = im_col.ee_collection.first()
# im_ee = im_ee.clip(region)
# im_ee = im_ee.select(['B4', 'B3', 'B2'])
# # Convert to xarray.Dataset
# im_xr = im_ee.wx.to_xarray(scale=30, crs='EPSG:4326')
# im_xr = xr.where(im_xr==im_xr.attrs['_FillValue'], np.nan, im_xr / 1e4)
# im_xr = im_xr.rio.write_crs('EPSG:4326')
# print('Sentinel-2 SR image loaded')

# # Reproject AOI and images to optimal UTM zone
# epsg_utm = convert_wgs_to_utm(aoi.geometry[0].centroid.coords.xy[0][0], aoi.geometry[0].centroid.coords.xy[1][0])
# aoi_utm = aoi.to_crs(f'EPSG:{epsg_utm}')
# im_xr = im_xr.rio.reproject(f'EPSG:{epsg_utm}')
# im_classified = im_classified.rio.write_crs("EPSG:4326")
# im_classified = im_classified.rio.reproject(f'EPSG:{epsg_utm}')
# im_classified = xr.where(im_classified < 1, np.nan, im_classified)

# Set up figure
fontsize=14
plt.rcParams.update({
    "font.size": fontsize,
    "font.sans-serif": "Arial",
    # "font.family": "sans-serif",
    # "font.sans-serif": "Computer Modern Sans Serif",
    "text.usetex": False
})

fig = plt.figure(figsize=(12,12))
gs = matplotlib.gridspec.GridSpec(2, 2, figure=fig, height_ratios=[1,2])
ax = [fig.add_subplot(gs[2:]),
      fig.add_subplot(gs[0]),
      fig.add_subplot(gs[1])]

# ----- Study sites
# GTOPO hillshade
ls = matplotlib.colors.LightSource(azdeg=90, altdeg=45)
ax[0].imshow(ls.hillshade(gtopo.data[0], vert_exag=0.002), cmap='gray', alpha=0.5,
             extent=(np.min(gtopo.x.data), np.max(gtopo.x.data), 
                     np.min(gtopo.y.data), np.max(gtopo.y.data)))
# RGI O2 region outlines
color = '#525252'
rgi_O2.plot(ax=ax[0], alpha=1.0, facecolor='None', edgecolor=color, linewidth=1)
ax[0].set_yticks(np.linspace(45, 65, num=6))
ax[0].set_xlim(-167, -112)
ax[0].set_ylim(46, 67)
ax[0].set_xlabel('Longitude')
ax[0].set_ylabel('Latitude')
ax[0].set_aspect(2.2)
# Median AARs
sns.scatterplot(data=min_snow_cover_stats, x='CenLon', y='CenLat', edgecolor='w', linewidth=0.5, 
                hue='Subregion', hue_order=subregion_order, palette=dict(min_snow_cover_stats[['Subregion', 'color']].drop_duplicates().values), 
                alpha=1, size='AAR_P50_min', sizes=(2,100), ax=ax[0])
handles, labels = ax[0].get_legend_handles_labels()
Ikeep = np.argwhere(['0.' in x for x in np.array(labels)]).flatten()
handles, labels = [handles[i] for i in Ikeep], [labels[i] for i in Ikeep] 
ax[0].legend(handles, labels, loc='lower left', title='2013–2023 median AAR', bbox_to_anchor=[0.2, 0.05, 0.2, 0.2])
# Add region labels and arrows
fontweight = 'bold'
ax[0].text(-163, 56, 'Aleutians', color=color, rotation=35, fontsize=fontsize-1, fontweight=fontweight)
ax[0].text(-158, 62.3, 'Alaska Range', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax[0].text(-147.9, 57.8, 'W. Chugach \nMtns.', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax[0].arrow(-147.6, 58.8, 0, 0.8, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax[0].text(-141.7, 57.7, 'St. Elias \nMtns.', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax[0].arrow(-141.5, 58.7, 0, 0.8, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax[0].text(-139.6, 56.4, 'N. Coast \nRanges', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax[0].arrow(-137.3, 56.8, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax[0].text(-133, 51.3, 'N. Cascades', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax[0].arrow(-129.4, 51.4, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax[0].text(-129.7, 47, 'S. Cascades', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax[0].arrow(-126, 47.1, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax[0].text(-132, 64, 'N. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax[0].text(-122, 55, 'C. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax[0].text(-117.7, 47, 'S. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
# Example site location
min_snow_cover_stats_site = min_snow_cover_stats.loc[min_snow_cover_stats['RGIId']==site_name]
ax[0].plot(min_snow_cover_stats_site['CenLon'], min_snow_cover_stats_site['CenLat'], '*', 
           markeredgecolor='k', markerfacecolor='#e7298a', markersize=15, linewidth=2)
    
# -----b) Sentinel-2 image
ax[1].imshow(np.dstack([im_xr.B4.data[0], im_xr.B3.data[0], im_xr.B2.data[0]]),
             extent=(np.min(im_xr.x.data)/1e3, np.max(im_xr.x.data)/1e3, 
                     np.min(im_xr.y.data)/1e3, np.max(im_xr.y.data)/1e3))
ax[1].set_xlabel('Easting [km]')
ax[1].set_ylabel('Northing [km]')

# -----c) Classified image
ax[2].imshow(im_classified.classified.data[0], cmap=cmap_classified, clim=(1,5),
             extent=(np.min(im_classified.x.data)/1e3, np.max(im_classified.x.data)/1e3,
                     np.min(im_classified.y.data)/1e3, np.max(im_classified.y.data)/1e3))
ax[2].plot(np.divide(aoi_utm.geometry[0].exterior.coords.xy[0], 1e3), np.divide(aoi_utm.geometry[0].exterior.coords.xy[1], 1e3),
           '-k', label='Glacier boundary')
ax[2].set_xlabel('Easting [km]')
# dummy points for legend
ax[2].plot(0, 0, 's', color=cmap_classified(0), label='Snow')
ax[2].plot(0, 0, 's', color=cmap_classified(1), label='Shadowed snow')
ax[2].plot(0, 0, 's', color=cmap_classified(2), label='Ice/firn')
ax[2].plot(0, 0, 's', color=cmap_classified(3), label='Rock/debris')
ax[2].plot(0, 0, 's', color=cmap_classified(4), label='Water')
ax[2].set_xlim(ax[1].get_xlim())
ax[2].set_ylim(ax[1].get_ylim())
handles, labels = ax[2].get_legend_handles_labels()
fig.legend(handles, labels, loc='upper center', ncols=6, markerscale=2, frameon=False)

# Plot AOI
for axis in ax[1:]:
    axis.plot(np.divide(aoi_utm.geometry[0].exterior.coords.xy[0], 1e3), 
              np.divide(aoi_utm.geometry[0].exterior.coords.xy[1], 1e3),
              '-k', linewidth=1, label='Glacier boundary')
    axis.set_yticks(np.arange(7030, 7046, step=5))

# Add text labels
text_labels = ['c', 'a', 'b']
for i in range(0, len(ax)):
    ax[i].text((ax[i].get_xlim()[1] - ax[i].get_xlim()[0]) * 0.9 + ax[i].get_xlim()[0],
               (ax[i].get_ylim()[1] - ax[i].get_ylim()[0]) * 0.85 + ax[i].get_ylim()[0],
                text_labels[i], fontweight='bold', fontsize=fontsize+4, horizontalalignment='center',
              bbox=dict(facecolor='w', edgecolor='None', pad=3))

# # Add caption
# caption = (r"\noindent\textbf{Figure 1. a)} Sentinel-2 surface reflectance image captured 2023-08-02 for one glacier (Randolph Glacier Inventory ID = 1.00037) \\"
#             r"and the associated \textbf{b)} classified image generated from the automated snow detection pipeline. \textbf{c)} Map of the study glacier locations, \\"
#             r"with marker sizes indicating the median accumulation area ratio (AAR) for the 2013–2023 study period. The maroon start marks the \\"
#             r"location of the example glacier shown in panels \textbf{a} and \textbf{b}." )
# fig.text(0.05, -0.02, caption, ha='left', wrap=True, fontsize=fontsize+1)

fig.tight_layout()
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'agu24_abstract_figure.png')
save_figure(fig, fig_fn)

### Snow cover GIF

import glob

# Load inputs
rgi_id = "RGI60-01.00312"
aoi_fn = os.path.join(scm_path, 'study-sites', rgi_id, 'AOIs', f"{rgi_id}_outline.shp")
aoi = gpd.read_file(aoi_fn)
im_classified_fns = sorted(glob.glob(os.path.join(scm_path, 'study-sites', rgi_id, 'imagery', 'classified', '*.nc')))
scs_fn = os.path.join(scm_path, 'study-sites', rgi_id, f"{rgi_id}_snow_cover_stats.csv")
scs = pd.read_csv(scs_fn)
scs['datetime'] = pd.to_datetime(scs['datetime'], format='mixed')
# subset to 2019
im_classified_fns = [x for x in im_classified_fns if int(os.path.basename(x)[0:4]) == 2019] 
scs = scs.loc[scs['datetime'].dt.year == 2019]
out_path = os.path.join(figures_out_path, 'timeseries_gif')
if not os.path.exists(out_path):
    os.mkdir(out_path)
    print('Made directory for outputs:', out_path)

# Define colormap for classified images
cmap_dict = {"Snow": "#4eb3d3",  "Shadowed_snow": "#636363", "Ice": "#084081", "Rock": "#fe9929", "Water": "#252525"}
colors = []
for key in list(cmap_dict.keys()):
    color = list(matplotlib.colors.to_rgb(cmap_dict[key]))
    if key=='Rock':
        color += [0.5]
    colors.append(color)
        
cmap = matplotlib.colors.ListedColormap(colors)
cmap

### Download images for 2019 ###

import math
import ee
import geedim as gd
import datetime
from rasterio.features import geometry_mask
    
def query_gee_for_imagery_run_pipeline(dataset, aoi_utm, date_start, date_end,
                                       month_start, month_end, site_name, 
                                       mask_clouds=True, cloud_cover_max=70, aoi_coverage=70, im_out_path=None,
                                       verbose=False, im_download=False):

    # -----Grab optimal UTM zone from AOI CRS
    epsg_utm = str(aoi_utm.crs.to_epsg())

    # -----Reformat AOI for image filtering
    # reproject CRS from AOI to WGS
    aoi_wgs = aoi_utm.to_crs('EPSG:4326')
    # prepare AOI for querying geedim (AOI bounding box)
    region = {'type': 'Polygon',
              'coordinates': [[[aoi_wgs.geometry.bounds.minx[0], aoi_wgs.geometry.bounds.miny[0]],
                               [aoi_wgs.geometry.bounds.maxx[0], aoi_wgs.geometry.bounds.miny[0]],
                               [aoi_wgs.geometry.bounds.maxx[0], aoi_wgs.geometry.bounds.maxy[0]],
                               [aoi_wgs.geometry.bounds.minx[0], aoi_wgs.geometry.bounds.maxy[0]],
                               [aoi_wgs.geometry.bounds.minx[0], aoi_wgs.geometry.bounds.miny[0]]
                               ]]}

    # -----Define function to query GEE for imagery
    def query_gee(dataset, date_start, date_end, region, cloud_cover_max, mask_clouds):
        if dataset == 'Landsat8':
            # Landsat 8
            im_col_gd = gd.MaskedCollection.from_name('LANDSAT/LC08/C02/T1_L2').search(start_date=date_start,
                                                                                       end_date=date_end,
                                                                                       region=region,
                                                                                       cloudless_portion=100 - cloud_cover_max,
                                                                                       mask=mask_clouds)
        elif dataset == 'Landsat9':
            # Landsat 9
            im_col_gd = gd.MaskedCollection.from_name('LANDSAT/LC09/C02/T1_L2').search(start_date=date_start,
                                                                                       end_date=date_end,
                                                                                       region=region,
                                                                                       cloudless_portion=100 - cloud_cover_max,
                                                                                       mask=mask_clouds)
        elif dataset == 'Sentinel-2_TOA':
            im_col_gd = gd.MaskedCollection.from_name('COPERNICUS/S2_HARMONIZED').search(start_date=date_start,
                                                                                         end_date=date_end,
                                                                                         region=region,
                                                                                         cloudless_portion=100 - cloud_cover_max,
                                                                                         mask=mask_clouds)

        elif dataset == 'Sentinel-2_SR':
            im_col_gd = gd.MaskedCollection.from_name('COPERNICUS/S2_SR_HARMONIZED').search(start_date=date_start,
                                                                                            end_date=date_end,
                                                                                            region=region,
                                                                                            cloudless_portion=100 - cloud_cover_max,
                                                                                            mask=mask_clouds)
        else:
            print("'dataset' variable not recognized. Please set to 'Landsat', 'Sentinel-2_TOA', or 'Sentinel-2_SR'. "
                  "Exiting...")
            return 'N/A'

        return im_col_gd

    # -----Define function to filter image IDs by month range
    def filter_im_ids_month_range(im_ids, im_dts, month_start, month_end):
        i = [int(ii) for ii in np.arange(0, len(im_dts)) if
             (im_dts[ii].month >= month_start) and (im_dts[ii].month <= month_end)]  # indices of images to keep
        im_ids, im_dts = [im_ids[ii] for ii in i], [im_dts[ii] for ii in i]  # subset of image IDs and datetimes
        # return 'N/A' if no images remain after filtering by month range
        if len(im_dts) < 1:
            return 'N/A', 'N/A'
        return im_ids, im_dts

    # -----Define function to couple image IDs captured within the same hour for mosaicking
    def image_mosaic_ids(im_col_gd):
        # Grab image properties, IDs, and datetimes from image collection
        properties = im_col_gd.properties
        ims = dict(properties).keys()
        im_ids = [properties[im]['system:id'] for im in ims]
        # return if no images found
        if len(im_ids) < 1:
            return 'N/A', 'N/A'
        im_dts = np.array(
            [datetime.datetime.utcfromtimestamp(properties[im]['system:time_start'] / 1000) for im in ims])

        # Remove image datetimes and IDs outside the specified month range
        im_ids, im_dts = filter_im_ids_month_range(im_ids, im_dts, month_start, month_end)

        # Grab all unique hours in image datetimes
        hours = np.array(im_dts, dtype='datetime64[h]')
        unique_hours = sorted(set(hours))

        # Create list of IDs for each unique hour
        im_mosaic_ids_list, im_mosaic_dts_list = [], []
        for unique_hour in unique_hours:
            i = list(np.ravel(np.argwhere(hours == unique_hour)))
            im_ids_list_hour = [im_ids[ii] for ii in i]
            im_mosaic_ids_list.append(im_ids_list_hour)
            im_dts_list_hour = [im_dts[ii] for ii in i]
            im_mosaic_dts_list.append(im_dts_list_hour)

        return im_mosaic_ids_list, im_mosaic_dts_list

    # -----Define function for extracting valid image IDs
    def extract_valid_image_ids(ds, date_start, date_end, region, cloud_cover_max, mask_clouds):
        # Initialize list of date ranges for querying
        date_ranges = [(date_start, date_end)]
        # Initialize list of error dates
        error_dates = []
        # Initialize error flag
        error_occurred = True
        # Iterate until no errors occur
        while error_occurred:
            error_occurred = False  # Reset the error flag at the beginning of each iteration
            try:
                # Initialize list of image collections
                im_col_gd_list = []
                # Iterate over date ranges
                for date_range in date_ranges:
                    # Query GEE for imagery
                    im_col_gd = query_gee(ds, date_range[0], date_range[1], region, cloud_cover_max, mask_clouds)
                    properties = im_col_gd.properties  # Error will occur here if an image is inaccessible!
                    im_col_gd_list.append(im_col_gd)
                # Initialize list of filtered image IDs and datetimes
                im_mosaic_ids_list_full, im_mosaic_dts_list_full = [], []  # Initialize lists of
                # Filter image IDs for month range and couple IDs for mosaicking
                for im_col_gd in im_col_gd_list:
                    im_mosaic_ids_list, im_mosaic_dts_list = image_mosaic_ids(im_col_gd)
                    if type(im_mosaic_ids_list) is str:
                        return 'N/A', 'N/A'
                    # append to list
                    im_mosaic_ids_list_full = im_mosaic_ids_list_full + im_mosaic_ids_list
                    im_mosaic_dts_list_full = im_mosaic_dts_list_full + im_mosaic_dts_list

                return im_mosaic_ids_list_full, im_mosaic_dts_list_full

            except Exception as e:
                error_id = str(e).split('ID=')[1].split(')')[0]
                print(f"Error querying GEE for {str(error_id)}")

                # Parse the error date from the exception message (replace this with your actual parsing logic)
                error_date = datetime.datetime.strptime(error_id[0:8], '%Y%m%d')
                error_dates.append(error_date)

                # Update date ranges excluding the problematic date
                date_starts = [date_start] + [str(error_date + datetime.timedelta(days=1))[0:10] for error_date in
                                              error_dates]
                date_ends = [str(error_date - datetime.timedelta(days=1))[0:10] for error_date in error_dates] + [
                    date_end]
                date_ranges = list(zip(date_starts, date_ends))

                # Set the error flag to indicate that an error occurred
                error_occurred = True

    # -----Apply functions
    if dataset == 'Landsat':  # must run Landsat 8 and 9 separately
        im_ids_list_8, im_dts_list_8 = extract_valid_image_ids('Landsat8', date_start, date_end, region,
                                                               cloud_cover_max, mask_clouds)
        im_ids_list_9, im_dts_list_9 = extract_valid_image_ids('Landsat9', date_start, date_end, region,
                                                               cloud_cover_max, mask_clouds)
        if (type(im_ids_list_8) is str) and (type(im_ids_list_9) is str):
            im_ids_list, im_dts_list = 'N/A', 'N/A'
        elif type(im_ids_list_9) is str:
            im_ids_list, im_dts_list = im_ids_list_8, im_dts_list_8
        elif type(im_ids_list_8) is str:
            im_ids_list, im_dts_list = im_ids_list_9, im_dts_list_9
        else:
            im_ids_list = im_ids_list_8 + im_ids_list_9
            im_dts_list = im_dts_list_8 + im_dts_list_9
    else:
        im_ids_list, im_dts_list = extract_valid_image_ids(dataset, date_start, date_end, region, cloud_cover_max,
                                                           mask_clouds)

    # -----Check if any images were found after filtering
    if type(im_ids_list) is str:
        print('No images found or error in one or more image IDs, exiting...')
        return 'N/A'

    if dataset=='Landsat':
        res = 30
        image_scalar = 36363.63636363636
        no_data_value = 0
    elif 'Sentinel-2' in dataset:
        res = 10
        image_scalar = 1e4
        no_data_value = -9999

    # -----Create xarray.Datasets from list of image IDs
    # loop through image IDs
    for i in tqdm(range(0, len(im_ids_list))):

        # subset image IDs and image datetimes
        im_ids, im_dts = im_ids_list[i], im_dts_list[i]

        # make directory for outputs (out_path) if it doesn't exist
        if not os.path.exists(im_out_path):
            os.mkdir(im_out_path)
            print('Made directory for image downloads: ' + im_out_path)
            
        # define filename
        if len(im_dts) > 1:
            im_fn = dataset + '_' + str(im_dts[0]).replace('-', '')[0:8] + '_MOSAIC.tif'
        else:
            im_fn = dataset + '_' + str(im_dts[0]).replace('-', '')[0:8] + '.tif'
            
        # check file does not already exist in directory, download
        if not os.path.exists(os.path.join(im_out_path, im_fn)):
            # create list of MaskedImages from IDs
            im_gd_list = [gd.MaskedImage.from_id(im_id) for im_id in im_ids]
            # combine into new MaskedCollection
            im_collection = gd.MaskedCollection.from_list(im_gd_list)
            # create image composite
            im_composite = im_collection.composite(method=gd.CompositeMethod.q_mosaic,
                                                   mask=mask_clouds,
                                                   region=region)
            # download to file
            im_composite.download(os.path.join(im_out_path, im_fn),
                                  region=region,
                                  scale=res,
                                  crs='EPSG:4326',
                                  dtype='int16',
                                  bands=im_composite.refl_bands)

    return

# Reproject AOI to optimal UTM zone
epsg_utm = f.convert_wgs_to_utm(aoi.geometry[0].centroid.coords.xy[0][0], aoi.geometry[0].centroid.coords.xy[1][0])
aoi_utm = aoi.to_crs(epsg_utm)

ee.Initialize()

# Landsat
dataset = 'Landsat'
query_gee_for_imagery_run_pipeline(dataset, aoi_utm, "2019-01-01", "2019-12-01", 5, 11, rgi_id, 
                                   mask_clouds=True, cloud_cover_max=70, aoi_coverage=70, im_out_path=out_path,
                                   verbose=False, im_download=True)

# Sentinel-2 SR
dataset = 'Sentinel-2_SR'
query_gee_for_imagery_run_pipeline(dataset, aoi_utm, "2019-01-01", "2019-12-01", 5, 11, rgi_id, 
                                   mask_clouds=True, cloud_cover_max=70, aoi_coverage=70, im_out_path=out_path,
                                   verbose=False, im_download=True)

# Sentinel-2 SR
dataset = 'Sentinel-2_TOA'
query_gee_for_imagery_run_pipeline(dataset, aoi_utm, "2019-01-01", "2019-12-01", 5, 11, rgi_id, 
                                   mask_clouds=True, cloud_cover_max=70, aoi_coverage=70, im_out_path=out_path,
                                   verbose=False, im_download=True)

# Remove images with < 70% coverage of AOI
fns = sorted(glob.glob(os.path.join(out_path, '*.tif')))
fn_remove_list = []
for fn in tqdm(fns):
    im = rxr.open_rasterio(fn).squeeze()
    im = im.rio.reproject(epsg_utm)
    im = im.isel(band=0)
    mask = geometry_mask(aoi_utm.geometry, transform=im.rio.transform(), invert=True, out_shape=im.shape)
    im_masked = xr.where(mask==1, im, np.nan)
    im_masked = xr.where((im_masked==-9999) | (im_masked==-32768), np.nan, im_masked)
    nreal = im_masked.notnull().sum().item()
    npx = np.count_nonzero(mask)
    percentage_real = (nreal / npx) * 100
    if percentage_real < 70:
        fn_remove_list.append(fn)
# for fn in fn_remove_list:
#     os.remove(fn)

# Iterate over classified images
for im_classified_fn in tqdm(im_classified_fns):
    # Load classified image
    im_classified = rxr.open_rasterio(im_classified_fn, decode_times=False).squeeze()
    im_classified = xr.where(im_classified==-9999, np.nan, im_classified)
    im_classified = xr.where(im_classified==2, 1, im_classified)
    date = pd.Timestamp(datetime.datetime.strptime(os.path.basename(im_classified_fn).split('_')[0], "%Y%m%dT%H%M%S"))
    source = os.path.basename(im_classified_fn).split(rgi_id + '_')[1].split('_classified.nc')[0]

    # Load multispec image
    try:
        im_fn = glob.glob(os.path.join(out_path, f"{source}_{str(date)[0:10].replace('-', '')}*.tif"))[0]
    except:
        continue
    im = rxr.open_rasterio(im_fn).squeeze()
    if source=='Landsat':
        image_scaler = 36363.63636363636
    else:
        image_scaler = 1e4
    im = xr.where((im==-9999) | (im==-32768), np.nan, im / image_scaler)
    
    # Plot
    fig = plt.figure(figsize=(8,8))
    gs = matplotlib.gridspec.GridSpec(2, 2, figure=fig, height_ratios=[2,1])
    ax = [fig.add_subplot(gs[0,0]), fig.add_subplot(gs[0,1]), fig.add_subplot(gs[1,:])]
    # RGB image
    ax[0].imshow(np.dstack([im.isel(band=2).data, im.isel(band=1).data, im.isel(band=0).data]),
                 extent=(np.min(im.x.data), np.max(im.x.data), np.min(im.y.data), np.max(im.y.data)))
    aoi.plot(ax=ax[0], facecolor='None', edgecolor='k')
    # classified image
    xmin, xmax = -145.3555, -145.045
    ymin, ymax = 63.144, 63.325
    ax[1].imshow(im_classified.data, cmap=cmap, clim=(1,5), 
                 extent=(xmin, xmax, ymax, ymin))
    ax[1].invert_yaxis()
    aoi.plot(ax=ax[1], facecolor='None', edgecolor='k')
    # dummy points for legend
    ax[1].plot(0, 0, 's', color=colors[0], markersize=12, label='Snow')
    ax[1].plot(0, 0, 's', color=colors[2], markersize=12, label='Ice')
    ax[1].plot(0, 0, 's', color=colors[3], markersize=12, label='Rock')
    ax[1].plot(0, 0, 's', color=colors[4], markersize=12, label='Water')
    ax[1].legend(loc='lower right', frameon=False)
    for axis in ax[0:2]:
        axis.set_xlim(xmin, xmax)
        axis.set_ylim(ymin, ymax)
        axis.set_xticks(axis.get_xticks()[1::2])
        axis.set_yticks(axis.get_yticks()[1::2])

    # AAR time series
    ax[2].plot(scs['datetime'], scs['AAR'], '.k')
    scs_date = scs.loc[(scs['datetime']==date) & (scs['source']==source)]
    ax[2].plot(scs_date['datetime'], scs_date['AAR'], '*m', markersize=15)
    ax[2].set_ylim(0,1.05)
    ax[2].grid(True)
    ax[2].set_ylabel('Snow area ratio')
    fig.suptitle(f"{date}\n{source.replace('_', ' ')}")
    
    fig.tight_layout()

    fig_fn = os.path.join(out_path, f"{date}_{rgi_id}_snow_cover.png")
    save_figure(fig, fig_fn)
    plt.close()
        
    

### Map of study sites with climate clusters

# Load climate clusters / mean climate
clusters = pd.read_csv(clusters_fn)
print('Climate clusters loaded')

# Load AOIs
aois = gpd.read_file(aois_fn)
# Add climate cluster column
aois = aois.merge(clusters[['RGIId', 'clustName']], on='RGIId')
aois.rename(columns={'clustName': 'Climate class'}, inplace=True)
print('AOIs loaded')

# Load RGI O2 Regions
rgi_O2_fn = os.path.join(scm_path, '..', 'GIS_data', 'RGI', 'RGIv7_02Regions', 
                                'RGI2000-v7.0-o2regions-Alaska-westernCanadaUS_clipped_to_country_outlines.shp')
rgi_O2 = gpd.read_file(rgi_O2_fn)
# remove Brooks Range
rgi_O2 = rgi_O2.loc[rgi_O2['o2region']!='01-01']
# add subregion name and color column
rgi_O2[['Subregion', 'color']] = '', ''
for i, o1o2 in enumerate(rgi_O2['o2region'].drop_duplicates().values):
    o1 = int(o1o2[0:2])
    o2 = int(o1o2[3:])
    subregion_name, color = f.determine_subregion_name_color(o1, o2)
    rgi_O2.loc[rgi_O2['o2region']==o1o2, 'Subregion'] = subregion_name
print('RGI O2 regions loaded')

# Load GTOPO30
gtopo_fn = '/Users/raineyaberle/Research/PhD/GIS_data/GTOPO30_clip.tif'
gtopo = rxr.open_rasterio(gtopo_fn)
gtopo = xr.where(gtopo==-32768, np.nan, gtopo)
print('GTOPO30 loaded')


# Set up figure
fontsize=12
plt.rcParams.update({'font.size':fontsize, 'font.sans-serif':'Arial'})
fig, ax = plt.subplots(figsize=(10,10))

# GTOPO hillshade
ls = matplotlib.colors.LightSource(azdeg=90, altdeg=45)
ax.imshow(ls.hillshade(gtopo.data[0], vert_exag=0.002), cmap='gray', alpha=0.5,
             extent=(np.min(gtopo.x.data), np.max(gtopo.x.data), 
                     np.min(gtopo.y.data), np.max(gtopo.y.data)))
# RGI O2 region outlines
color = '#525252'
rgi_O2.plot(ax=ax, alpha=1.0, facecolor='None', edgecolor=color, linewidth=1)
ax.set_yticks(np.linspace(45, 65, num=6))
ax.set_xlim(-167, -112)
ax.set_ylim(46, 66.5)
ax.set_xlabel('Longitude')
ax.set_ylabel('Latitude')
ax.set_aspect(2.1)
# Site locations
obj = sns.scatterplot(data=aois, x='CenLon', y='CenLat', edgecolor='k', linewidth=0.5, 
                      hue='Climate class', hue_order=cluster_order, palette=cluster_cmap_dict, alpha=1, ax=ax)
sns.move_legend(obj, loc='lower left', markerscale=2, bbox_to_anchor=[0.15, 0.1, 0.2, 0.2])
# Add region labels and arrows
fontweight = 'bold'
ax.text(-163, 56, 'Aleutians', color=color, rotation=35, fontsize=fontsize-1, fontweight=fontweight)
ax.text(-158, 62.3, 'Alaska Range', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax.text(-147.9, 57.8, 'W. Chugach \nMtns.', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax.arrow(-147.6, 58.8, 0, 0.8, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax.text(-141.7, 57.7, 'St. Elias \nMtns.', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax.arrow(-141.5, 58.7, 0, 0.8, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax.text(-139.6, 56.4, 'N. Coast \nRanges', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax.arrow(-137.3, 56.8, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax.text(-133, 51.3, 'N. Cascades', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax.arrow(-129.4, 51.4, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax.text(-129.7, 47, 'S. Cascades', color=color, rotation=0, horizontalalignment='center', fontsize=fontsize-1, fontweight=fontweight)
ax.arrow(-126, 47.1, 1.3, 0, color=color, linewidth=2, head_width=0.25, head_length=0.2)
ax.text(-132, 64, 'N. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax.text(-122, 55, 'C. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)
ax.text(-117.7, 47, 'S. Rockies', color=color, rotation=0, fontsize=fontsize-1, fontweight=fontweight)

plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'agu24_study_sites_map.png')
save_figure(fig, fig_fn)


### Model SLA animation

import glob
plt.rcParams.update({'font.size':12, 'font.sans-serif': 'Arial'})

# Define output directory
out_path = os.path.join(figures_out_path, 'model_SMB_to_SLA_animation')
if not os.path.exists(out_path):
    os.mkdir(out_path)
    print('Made directory for outputs:', out_path)
    
# Grab modeled SMB file names
rgi_id = '1.00032'
bin_fn = sorted(glob.glob(os.path.join(scm_path, 'Rounce_et_al_2023', 'binned', f'{rgi_id}*.nc')))[0]

# Load binned data
bin = xr.open_dataset(bin_fn)
# grab data variables
h = bin.bin_surface_h_initial.data[0] # surface elevation [m]
x = bin.bin_distance.data[0]
b_sum = np.zeros((len(bin.time.data), len(h))) # cumulative SMB
times = [np.datetime64(x) for x in bin.time.data] # datetimes
months = list(pd.DatetimeIndex(times).month) # months of each datetime
# iterate over each time period after 2013
times = [time for time in times if time >= np.datetime64('2012-10-01')]
elas = np.nan*np.zeros(len(times)) # initialize transient ELAs
for j, time in enumerate(tqdm(times)):
    # subset binned data to time
    bin_time = bin.isel(time=j)
    # grab the SMB 
    b_sum[j,:] = bin_time.bin_massbalclim_monthly.data[0]
    # add the previous SMB (restart the count in October)
    if months[j] != 10: 
        b_sum[j,:] += b_sum[j-1,:]
    # If all SMB > 0, ELA = minimum elevation
    if all(b_sum[j,:] > 0):
        elas[j] = np.min(h)
    # If SMB is > 0 and < 0 in some places, linearly interpolate ELA
    elif any(b_sum[j,:] < 0) & any(b_sum[j,:] > 0):
        elas[j] = np.interp(0, np.flip(b_sum[j,:]), np.flip(h))
    # If SMB < 0 everywhere, fit a piecewise linear fit and extrapolate for SMB=0
    elif all(b_sum[j,:] < 0):
        X, y = b_sum[j,:], h
        elas[j] = np.nanmax(h)
    else:
        print('issue')
        
    # Plot results
    if time >= np.datetime64('2013-01-01'):
        fig, ax = plt.subplots(2, 1, gridspec_kw=dict(height_ratios=[3,1]), figsize=(6,6))
        # surface profile
        ax[0].fill_between(x, np.zeros(len(x)), h, color='gray', edgecolor='k', alpha=0.5)
        positive_mask = b_sum[j,:] > 0
        negative_mask = b_sum[j,:] < 0
        # positive mass balance bars
        ax[0].bar(x[positive_mask], b_sum[j,positive_mask]*50, width=126, 
            bottom=h[positive_mask], color='blue', alpha=0.6, label='Positive Mass Balance')
        # negative mass balance bars
        ax[0].bar(x[negative_mask], b_sum[j,negative_mask]*50, width=126, 
                bottom=h[negative_mask], color='red', alpha=0.6, label='Negative Mass Balance')
        # SLA
        ax[0].axhline(elas[j], 0, np.nanmax(x), color='k')
        ax[0].text(7e3, elas[j]+50, 'Snowline altitude', color='k', ha='right')
        ax[0].set_ylim(1.25e3, 3.3e3)
        ax[0].set_yticks([1500, 2000, 2500, 3000])
        ax[0].set_ylabel('Elevation [m]')
        ax[0].set_xlim(0, np.nanmax(x))
        ax[0].set_xticks(np.arange(0, 7.1e3, step=1e3))
        ax[0].set_xticklabels(np.divide(ax[0].get_xticks(), 1e3).astype(int).astype(str))
        ax[0].set_xlabel('km')
        ax[0].spines[['top', 'right']].set_visible(False)
        ax[0].set_title(str(time)[0:7])
        # SLA time series
        ax[1].plot(times, elas, '.k', markersize=5)
        ax[1].plot(times[j], elas[j], '*m', markersize=10)
        ax[1].set_xlim(np.datetime64('2013-01-01'), np.datetime64('2023-01-01'))
        ax[1].set_ylim(1330, 2600)
        ax[1].set_ylabel('Snowline altitude [m]')
        fig.tight_layout()
        # Save figure
        fig_fn = os.path.join(out_path, f"{str(time)[0:7]}.png")
        save_figure(fig, fig_fn)
        plt.close()
        

## Site location map

# # Load GTOPO30
# gtopo_fn = '/Users/raineyaberle/Research/PhD/GIS_data/GTOPO30_clip.tif'
# gtopo = rxr.open_rasterio(gtopo_fn)
# gtopo = xr.where(gtopo==-32768, np.nan, gtopo)
# print('GTOPO30 loaded')
# Load country outlines
countries_fn = '/Users/raineyaberle/Research/PhD/GIS_data/countries_shp/countries.shp'
countries = gpd.read_file(countries_fn)
countries = countries.loc[(countries['NAME']=='Canada') | (countries['NAME']=='United States')]

# Set up figure
fontsize=12
plt.rcParams.update({'font.size':fontsize, 'font.sans-serif':'Arial'})
fig, ax = plt.subplots(figsize=(8,8))
# GTOPO hillshade
ls = matplotlib.colors.LightSource(azdeg=90, altdeg=45)
ax.imshow(ls.hillshade(gtopo.data[0], vert_exag=0.002), cmap='gray', alpha=0.5,
             extent=(np.min(gtopo.x.data), np.max(gtopo.x.data), 
                     np.min(gtopo.y.data), np.max(gtopo.y.data)))
# country outlines
countries.plot(ax=ax, facecolor='None', edgecolor='gray')
# Glaciers
# aois.plot(facecolor='#08519c', edgecolor='None', ax=ax)
# Site location
rgi_id = 'RGI60-01.00032'
site_centroid = aois.loc[aois['RGIId']==rgi_id].geometry[0].centroid.coords.xy
ax.plot(*site_centroid, '*', markersize=40, markerfacecolor='m', markeredgecolor='k', linewidth=0.5)

ax.set_xlim(-170, -120)
ax.set_ylim(45, 71)
ax.set_aspect(2.1)
ax.set_axis_off()

# Save figure
fig_fn = os.path.join(figures_out_path, 'model_SMB_to_SLA_animation', 'site_location_map.png')
save_figure(fig, fig_fn)


## Grad Student Showcase 2025: AARs timings and magnitudes

# -----Load glacier boundaries
aois = gpd.read_file(aois_fn)
print('Glacier boundaries loaded')

# -----Load climate clusters
clusters = pd.read_csv(clusters_fn)

# -----Load median AARs for all sites
min_scs_fn = os.path.join(out_path, 'minimum_snow_cover_stats.csv')
min_scs = pd.read_csv(min_scs_fn)
# Add difference from September AAR
min_scs['AAR_Sept-obs'] = min_scs['AAR_Sept'] - min_scs['AAR_median']
# Add Subregion and climate cluster info
min_scs[['CenLon', 'CenLat', 'Subregion', 'clustName']] = 0, 0, '', ''
for rgi_id in min_scs['RGIId'].drop_duplicates().values:
    cenlon, cenlat, subregion = aois.loc[aois['RGIId']==rgi_id, ['CenLon', 'CenLat', 'Subregion']].values[0]
    cluster = clusters.loc[clusters['RGIId']==rgi_id, 'clustName'].values[0]
    min_scs.loc[min_scs['RGIId']==rgi_id, ['CenLon', 'CenLat', 'Subregion', 'clustName']] = cenlon, cenlat, subregion, cluster
# Sort by subregion order
min_scs['Subregion'] = pd.Categorical(min_scs['Subregion'], subregion_order)
min_scs.sort_values(by='Subregion', inplace=True)
print('Median AARs loaded')

fontsize=14
lw = 1.5
plt.rcParams.update({'font.size': fontsize, 'font.sans-serif':'Arial'})
gs = matplotlib.gridspec.GridSpec(10,2, wspace=0.1, hspace=0.0)
fig = plt.figure(figsize=(10,8))
ax = []

# Iterate over subregions
median_color = 'w'
fill_color = '#2477BF'
Iax = -1
for i, subregion in enumerate(min_scs['Subregion'].unique()):
    min_scs_subregion = min_scs.loc[min_scs['Subregion']==subregion]
    
    # a) AAR timings
    ax.append(fig.add_subplot(gs[i,0]))
    Iax += 1
    k = sns.kdeplot(min_scs_subregion['WOY_median'], vertical=False, color=fill_color, 
                    fill=True, edgecolor='k', linewidth=lw, alpha=1, ax=ax[Iax], zorder=2)
    median = min_scs_subregion['WOY_median'].median()
    ax[Iax].plot([median, median], [0, ax[Iax].get_ylim()[1]*0.9], '-', color='w', linewidth=lw+0.5, zorder=3)
    ax[Iax].plot([14,45], [0,0], '-', color='k', linewidth=2)
    ax[Iax].set_xlim(24,47)
    ax[Iax].set_ylabel('')
    ax[Iax].set_xticks([])
    ax[Iax].set_xlabel('')
    if i==9:
        ax[Iax].set_xticks([26, 31, 35, 39, 44])
        ax[Iax].set_xticklabels(['Jul', 'Aug', 'Sep', 'Oct', 'Nov'])
        ax[Iax].set_xlabel('Time of year')
    if i==0:
        ax[Iax].set_title('a) Snow minima timings', fontsize=fontsize+4, color='k')
        ax[Iax].legend(loc='upper center', frameon=False, bbox_to_anchor=[0.4, 1.6, 0.2, 0.2])
    ax[Iax].spines[['right', 'top', 'bottom']].set_visible(False)
    ax[Iax].set_ylabel(subregion, ha='right', va='center', color='k', rotation=0)
    ax[Iax].set_yticks([ax[Iax].get_ylim()[1]*0.5])
    ax[Iax].set_yticklabels([])

    # b) AAR magnitude differences
    ax.append(fig.add_subplot(gs[i,1]))
    Iax += 1
    sns.boxplot(data=min_scs_subregion, x='AAR_Sept-obs', y='Subregion', showfliers=False, ax=ax[Iax],
                boxprops=dict(edgecolor='k', linewidth=lw, facecolor=fill_color),
                medianprops=dict(color='w', linewidth=lw-0.5), 
                whiskerprops=dict(color='k', linewidth=lw), 
                capprops=dict(color='k', linewidth=lw))
    ax[Iax].set_xlim(-0.05,0.6)
    ax[Iax].set_ylim(i-0.5, i+0.5)  
    ax[Iax].spines[['right', 'top']].set_visible(False)
    if i > 0:
        ax[Iax].spines['top'].set_visible(True)
        ax[Iax].spines['top'].set_color('gray')
    handles, labels = ax[Iax].get_legend_handles_labels()
    ax[Iax].legend().remove()
    if i==0:
        ax[Iax].set_title('b) September-only $-$ Observed AARs', fontsize=fontsize+4, color='k')
        labels = ['Observed', 'September-only']
        ax[Iax].legend(handles, labels, loc='upper center', ncols=1, frameon=False, 
                       bbox_to_anchor=[0.9, 0.9, 0.2, 0.2])
    ax[Iax].set_ylabel('')
    ax[Iax].set_yticklabels([])
    if i < 9:
        ax[Iax].set_xticks([])
        ax[Iax].set_xlabel('')
        ax[Iax].spines['bottom'].set_color('gray')
    else:
        ax[Iax].set_xlabel('AAR difference')
            
plt.show()

# Save figure
fig_fn = os.path.join(figures_out_path, 'fig02_median_aars+timings_swapped_larger-text.png')
save_figure(fig, fig_fn)
